### 2024-12-30 #7619 @editorialbot
[PRE REVIEW]: Minterpy: multivariate polynomial interpolation in Python
**Submitting author:** @damar-wicaksono (<a href="http://orcid.org/0000-0001-8587-7730">Damar Wicaksono</a>)
**Repository:** https://github.com/minterpy-project/minterpy
**Branch with paper.md** (empty if default branch): dev-joss
**Version:** v0.3.0
