### 2024-12-30 #7619 @danielskatz
👋 @damar-wicaksono - thanks for your submission. Note that your paper is about 2400 words long, while JOSS requests papers be roughly 250-1000 words. Please remove content from your paper, perhaps following the [example paper](https://joss.readthedocs.io/en/latest/example_paper.html). You might want to point from the paper to details in the repo or documentation, rather than putting all the details in the paper.

In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect.

Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. You can also use `@editorialbot check repository` to check a number of things, include the number of words in the paper.

editorialbot commands need to be the first entry in a new comment.

When you are ready to proceed, please ping me.
