### 2024-12-31 af82f14
Shorten the paper
