+++ joss/paper.bib
@@ -243,14 +243,6 @@
   publisher = {The Open Journal},
 }
 
-@Misc{Procida2024,
-  author = {Procida, Daniele},
-  note   = {Retrieved September 26, 2024},
-  title  = {{Diátaxis} documentation framework},
-  year   = {2024},
-  url    = {https://diataxis.fr/},
-}
-
 @Article{Trefethen2017a,
   author  = {Trefethen, L. N.},
   journal = {Proceedings of the American Mathematical Society},
