+++ joss/paper.md
@@ -42,21 +42,19 @@ bibliography: paper.bib
 
 # Summary
 
-Interpolation is the bottleneck in numerous computational tasks,
-appearing in function approximation, curve fitting, numerical integration,
-surface level-set methods, differential geometry,
-spectral methods, surrogate-model-based optimization,
-sensitivity analysis, uncertainty quantification,
-and beyond.
+Interpolation is essential in various computational tasks,
+including function approximation, curve fitting, numerical integration,
+differential geometry,
+spectral methods, optimization,
+and uncertainty quantification.
 
 Minterpy is an open-source Python package designed
 for multivariate polynomial interpolation.
-It provides a means of approximating a wide range of functions
-using numerically stable and accurate interpolating polynomials.
-Some of its key features are:
+It provides stable and accurate interpolating polynomials for approximating a wide range of functions.
+Key features include:
 
 - Polynomial interpolation on properly selected nodes and
-  polynomial regression on given arbitrary scattered nodes.
+  regression on arbitrary nodes.
 - Differentiation and integration operations on the polynomials.
 - Addition, subtraction, and multiplication operations
   on the polynomials.
@@ -67,303 +65,187 @@ commonly associated with interpolation tasks.
 
 # Statement of need
 
-Polynomial interpolation can be traced back to the works of
-Newton, Lagrange, and others [@Meijering2002],
-and its significance in mathematics and computing is well-established
-[@Cools2002;@Hecht2018;@Xiu2009].
-As a means of approximating functions, global polynomials---where a single polynomial
-is defined over the entire domain---offer several advantages
-that facilitate common numerical operations, such as
-differentiation, integration, addition, subtraction, or multiplication [@Trefethen2019].
+Polynomial interpolation, rooted in the works of Newton and Lagrange [@Meijering2002],
+plays a vital role in mathematics and computing [@Cools2002;@Hecht2018;@Xiu2009].
+Global polynomials---defined over the entire domain---enable efficient numerical operations
+like differentiation, integration, addition, subtraction, or multiplication [@Trefethen2019].
 
-Polynomial interpolation is based on the principle that, in one dimension,
-one and only one polynomial $Q_{f, n}$ of degree $n$ can interpolate
+In one dimension, a unique polynomial $Q_{f, n}$ of degree $n$ interpolates
 a function $f: \Omega \mapsto \mathbb{R}$ in a bounded domain $\Omega$
-with $n + 1$ distinct _unisolvent interpolation nodes (points)_ $P_n$ such that
+at $n + 1$ distinct _unisolvent interpolation nodes_ $P_n$ satisfying
 $$
 Q_{f, n} (p_i) = f(p_i),\; \forall p_i \in P_n \subset \Omega,\; i = 0, \ldots, n.
 $$
 The Stone-Weierstrass Approximation Theorem further states that,
 in multiple dimensions, any continuous function on a bounded domain
 can be approximated by multivariate global polynomials [@Branges1959].
-The theorem, however, does not prescribe how to obtain such a polynomial.
-Furthermore, the theorem does not even require the sequence of polynomials $Q_{f, n}$
+The theorem, however, does not provide construction method;
+it does not even require the sequence of polynomials $Q_{f, n}$
 to coincide with $f$ _anywhere_ in the domain,
 while still having uniform convergence on the domain.
-In short, approximation and interpolation
-using global polynomials are two distinct problems that must be addressed simultaneously.
+Thus, approximation and interpolation
[…]
