### 2024-12-31 #7619 @damar-wicaksono
@danielskatz: Thank you very much for handling our submission. Following your suggestion, we have revised the manuscript and significantly reduced its number of words by removing unnecessary details; they can still be found in the online documentation. I hope that it's okay.

We also carefully reviewed the list of references and were unable to locate the missing DOI for some papers as flagged by the bot.

Thank you again!
