### 2024-12-31 #7619 @danielskatz
Thanks @damar-wicaksono.

And DOIs aren't needed for all references, just DOIs for references where there is a DOI...
