### 2025-01-19 #7619 @xuanxu
@matbesancon @sdesabbata would you be willing to review this paper for JOSS? I think your expertise with Python and Math makes both of you a good fit for this submission.
