### 2025-01-21 #7619 @xuanxu
@sixpearls @kenohori would you be willing to review this paper for JOSS? I think your expertise with Python and Math makes both of you a good fit for this submission.
