### 2025-01-21 #7702 @editorialbot
[REVIEW]: Minterpy: multivariate polynomial interpolation in Python
**Submitting author:** @damar-wicaksono (<a href="http://orcid.org/0000-0001-8587-7730">Damar Wicaksono</a>)
**Repository:** https://github.com/minterpy-project/minterpy
**Branch with paper.md** (empty if default branch): dev-joss
**Version:** v0.3.1
**Archive:** 10.14278/rodare.3725
