### 2025-01-21 #7702 @xuanxu
@sixpearls @kenohori – This is the review thread for the paper. All of our communications will happen here from now on.
