### 2025-02-24 #7702 @xuanxu
@kenohori @sixpearls Friendly ping
