### 2025-03-21 #7702 @kenohori
Hello!

I’ve reviewed the paper and the software and everything seems okay. The software works well and the tests passed on my machine (MacBook Pro x86). The paper has all the required elements as well.

My only real concern is that the paper seems to be written more for mathematicians and experts on interpolation than for practitioners (non-specialists) who might be interested in using minterpy for their own problems. In order to make it more approachable, I’d encourage the authors to incorporate some more practical information about how to apply the library and how it differs from other libraries/methods in a more practical sense.

Apart from that, I have some minor comments on the paper:
* Line 29: provide *a* construction method
* Line 33: better compared to what?
* Line 38: what is appropriate in this context?
* Lines 53-54: in a practical sense, what are the advantages of this approach?
* Figure 1: is the code for the tests in this figure available?

Finally, I’d encourage the authors to add some relevant links from the repository’s README to the library documentation. There’s already one link that probably should go there, but it seems to be broken as it links to a private GitLab repository.
