### 2025-03-24 minterpy-project/minterpy#10 @damar-wicaksono
[JOSS] Revise paper according to JOSS Review (Reviewer 2)
In response to feedback from the JOSS submission review ([https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2744691848](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2744691848)), the paper should be revised to include some information on the practical application of the package as well as a clearer comparison with other packages, highlighting the unique strengths and advantages of Minterpy from a practical perspective. This will help make the paper more appealing and useful to practitioners who may be interested in using Minterpy.

Furthermore, several minor points should be revised and clarified:

> - Line 29: provide _a_ construction method
> - Line 33: better compared to what?
> - Line 38: what is appropriate in this context?
> - Lines 53-54: in a practical sense, what are the advantages of this approach?
> - Figure 1: is the code for the tests in this figure available?
