### 2025-03-24 minterpy-project/minterpy#11 @damar-wicaksono
[JOSS] Revise README.md according to JOSS Review (Reviewer 2)
Following a feedback on the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2744691848), several additional relevant links from the README.md to the library documentation may be added.

Furthermore, double check for any broken links in the README.
