### 2025-03-24 minterpy-project/minterpy#12 @damar-wicaksono
[JOSS] Submission to JOSS
Following issue #8, Minterpy has been submitted to JOSS:

- The [pre-review process](https://github.com/openjournals/joss-reviews/issues/7619) has been concluded.
- The [review process](https://github.com/openjournals/joss-reviews/issues/7702) is ongoing.

Relevant issues related to the review process:
- #10
- #11
