### 2025-03-26 #7702 @damar-wicaksono
@sixpearls:

- The code used to create Figure 1 is archived in [RODARE](https://rodare.hzdr.de/record/3379). The link was supposed to be referenced to in the paper, but I missed it.
- Thank you for pointing the issue on README. There is an [open issue](https://github.com/minterpy-project/minterpy/issues/11) related to revising the README. The issues you raised will also be resolved later.
