### 2025-04-01 #7702 @sixpearls
Thanks! This is a very interesting paper and I'm interested in learning more. I had a few suggestions, questions, and requests for more information regarding some of the incomplete items on my checklist.

### Reproducibility

I am having trouble interpreting the command line options for the run scripts. What values were used for Figure 1 in the paper? If they aren't there already, I would recommend documenting this with the paper examples, or perhaps even including them like examples (with different optional requirements) in the main repository.

### Quality of writing (Software paper)

(quick proof reading catches)
line 64: "where where"
line 80: citation/explanation for Newton stability?
line 94-95: "x \in ..." rendering error for R^n?

### A statement of need (Software paper and/or documentation)

I think it would be helpful if you could more clearly state the primary purpose of this approach? I read the following, but please correct or confirm:
- the multi-index set paramaterization lends itself to "reducing curse of dimensionality" (by using p=2)
- primary use case is approximation not interpolation, so use [non-equally spaced grid](https://minterpy-project.github.io/minterpy/latest/_images/e3e96bcda98441dac4d453d5bbb9ffef67fd8e4e81024b86034d8084feebd0f9.png) to address Runge phenomona
- this representation (or perhaps transformation to Newton basis) allows easy implementation of closure under arithmetic and differentiation

Can you provide a brief/high-level summary statement of the references explaining the following:
- How global polynomials are misperceived as unstable, susceptible to Runge's phenomone, etc. (lines 34-36)?
- What about the grid reduces addresses Runge phenomona? (assuming that was a correct interpretation)
- The meaning of curse of dimensionality in this context and how Euclidian degree reduces it
