### 2025-04-02 minterpy-project/minterpy#13 @damar-wicaksono
[JOSS] Add additional details to the archived numerical experiment (Reviewer 1)
Following the JOSS submission review ([Reviewer 1](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2768399816)), an issue on reproducibility of the numerical experiment to produce Figure 1 of the paper has been raised.

Specifically:

> I am having trouble interpreting the command line options for the run scripts. What values were used for Figure 1 in the paper? If they aren't there already, I would recommend documenting this with the paper examples, or perhaps even including them like examples (with different optional requirements) in the main repository.

Clarify this point with the reviewer and add additional details to the archived codes in [RODARE](https://rodare.hzdr.de/record/3379) (a new version will then be added).
