### 2025-04-02 minterpy-project/minterpy#14 @damar-wicaksono
[JOSS] Revise typos and grammatical mistakes in the paper (Reviewer 1)
Following the JOSS submission review ([Reviewer 1](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2768399816)), several mistakes in the paper have been found.

Specifically:

> (quick proof reading catches)
> line 64: "where where"
> line 80: citation/explanation for Newton stability?
> line 94-95: "x \in ..." rendering error for R^n?

Double check the paper again for other mistakes and fix them.
