### 2025-04-02 minterpy-project/minterpy#15 @damar-wicaksono
[JOSS] Fix issues in README.md (Reviewer 1)
Following the JOSS submission review ([Reviewer 1](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2754583234)), an additional issue in the README has been found (see Issue #11 for others).

Specifically:

> Similarly, the prose before the `test_function` definition refers to the wrong function `x/sin(10*x)` instead of `x*sin(10*x)`. Please make this consistent.

Double check the README again and fix any issues found.
