### 2025-04-02 minterpy-project/minterpy#16 @damar-wicaksono
[JOSS] Provide additional clarification related to the statement of need (Reviewer 1)
Following the JOSS submission review ([Reviewer 1](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2768399816)), additional clarification regarding the statement of the need is required.

Specifically:

> I think it would be helpful if you could more clearly state the primary purpose of this approach? I read the following, but please correct or confirm:
>
> - the multi-index set paramaterization lends itself to "reducing curse of dimensionality" (by using p=2)
> - primary use case is approximation not interpolation, so use [non-equally spaced grid](https://minterpy-project.github.io/minterpy/latest/_images/e3e96bcda98441dac4d453d5bbb9ffef67fd8e4e81024b86034d8084feebd0f9.png) to address Runge phenomona
> - this representation (or perhaps transformation to Newton basis) allows easy implementation of closure under arithmetic and differentiation
>
> Can you provide a brief/high-level summary statement of the references explaining the following:
>
> - How global polynomials are misperceived as unstable, susceptible to Runge's phenomone, etc. (lines 34-36)?
> - What about the grid reduces addresses Runge phenomona? (assuming that was a correct interpretation)
> - The meaning of curse of dimensionality in this context and how Euclidian degree reduces it

First, clarify with the reviewer regarding the points above. Then, update the paper when/if necessary.
