### 2025-04-02 #7702 @damar-wicaksono
@sixpearls,

Thank you very much for pointing out these issues!

To keep this thread a bit tidy, I've created separate issues in our [GitHub repository](https://github.com/minterpy-project/minterpy/issues) to address the points raised during this review process, including your feedback. I'll attempt at resolving each issue in the coming days and will mention you directly in the relevant issue if I need any further clarification or confirmation from you. Thank you again!
