### 2025-04-07 05716ea
Revise statement of need and package overview
The Statement of need has been revised according
to the feedback provided by Reviewer 1 in the JOSS
submission review.
Additional references have been added.
The explanation about the notion of the curse of dimensionality
has been expanded in Section 3 (Package Overview).
Finally, found typos have been fixed.

This commit should resolve Issue #10, #14, and #16.
