+++ joss/paper.bib
@@ -281,4 +281,57 @@
   doi     = {10.1016/j.jat.2012.11.005},
 }
 
+@Book{Goldstine1977,
+  author    = {Goldstine, Herman H.},
+  publisher = {Springer New York},
+  title     = {A history of numerical analysis from the 16th through the 19th century},
+  year      = {1977},
+  isbn      = {9781468494723},
+  series    = {Studies in the History of Mathematics and Physical Sciences},
+  doi       = {10.1007/978-1-4684-9472-3},
+}
+
+@Article{Trefethen2016,
+  author    = {Trefethen, L. N.},
+  journal   = {Notices of the American Mathematical Society},
+  title     = {Inverse {Yogiisms}},
+  year      = {2016},
+  number    = {11},
+  pages     = {1281--1285},
+  volume    = {63},
+  doi       = {10.1090/noti1446},
+  publisher = {American Mathematical Society (AMS)},
+}
+
+@Article{Breuss2018,
+  author  = {Breuß, Michael and Kemm, Friedemann and Vogel, Oliver},
+  journal = {Kybernetika},
+  title   = {A numerical study of {Newton} interpolation with extremely high degrees},
+  year    = {2018},
+  pages   = {279--288},
+  doi     = {10.14736/kyb-2018-2-0279},
+}
+
+@Article{Reichel1990,
+  author  = {Reichel, Lothar},
+  journal = {BIT},
+  title   = {Newton interpolation at {Leja} points},
+  year    = {1990},
+  number  = {2},
+  pages   = {332--346},
+  volume  = {30},
+  doi     = {10.1007/bf02017352},
+}
+
+@Article{TalEzer1991,
+  author  = {Tal-Ezer, Hillel},
+  journal = {SIAM Journal on Scientific and Statistical Computing},
+  title   = {High degree polynomial interpolation in {Newton} form},
+  year    = {1991},
+  number  = {3},
+  pages   = {648--667},
+  volume  = {12},
+  doi     = {10.1137/0912034},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
