+++ joss/paper.md
@@ -65,58 +65,98 @@ commonly associated with interpolation tasks.
 
 # Statement of need
 
-Polynomial interpolation, rooted in the works of Newton and Lagrange [@Meijering2002],
-plays a vital role in mathematics and computing [@Cools2002;@Hecht2018;@Xiu2009].
-Global polynomials---defined over the entire domain---enable efficient numerical operations
-like differentiation, integration, addition, subtraction, or multiplication [@Trefethen2019].
+As a means of approximating functions, global polynomials---where a single
+polynomial is defined over the entire domain---offer several advantages.
+For sufficiently smooth functions, global polynomials can achieve high accuracy
+with a smaller number of data points (sampled over the entire domain) compared
+to local piecewise polynomials.
+Additionally, their relatively simple structure facilitates
+many common numerical operations.
+These operations include differentiation, integration, addition, subtraction,
+and multiplication [@Trefethen2019].
 
-In one dimension, a unique polynomial $Q_{f, n}$ of degree $n$ interpolates
-a function $f: \Omega \mapsto \mathbb{R}$ in a bounded domain $\Omega$
-at $n + 1$ distinct _unisolvent interpolation nodes_ $P_n$ satisfying
+The Stone-Weierstrass theorem establishes that any continuous function
+on a bounded domain in multiple dimensions can be approximated uniformly
+to arbitrary precision by multivariate global polynomials [@Branges1959].
+However, the theorem does not specify a concrete method
+for constructing such approximating polynomials.
+Various techniques can be employed to build approximating polynomials,
+including Taylor expansions and least square approximations.
+Minterpy focuses on constructing approximating global polynomials
+using one of the earliest and most established methods:
+interpolation [@Goldstine1977].
+
+Polynomial interpolation is based on the principle that, in one dimension,
+there exists a unique polynomial $Q_{f, n}$ of degree $n$
+that interpolates a function $f: \Omega \to \mathbb{R}$ in a bounded domain
+$\Omega$ with $n + 1$ _distinct (unisolvent[^unisolvent]) interpolation nodes
+(or points)_ $P_n$ such that 
 $$
 Q_{f, n} (p_i) = f(p_i),\; \forall p_i \in P_n \subset \Omega,\; i = 0, \ldots, n.
 $$
-The Stone-Weierstrass Approximation Theorem further states that,
-in multiple dimensions, any continuous function on a bounded domain
-can be approximated by multivariate global polynomials [@Branges1959].
-The theorem, however, does not provide construction method;
-it does not even require the sequence of polynomials $Q_{f, n}$
-to coincide with $f$ _anywhere_ in the domain,
-while still having uniform convergence on the domain.
-Thus, approximation and interpolation
-using global polynomials are distinct challenges.
+Polynomial interpolation has its roots in the works of Newton, Euler, Lagrange,
+and others [@Meijering2002],
+and its significance in mathematics and computing
+is well-established [@Cools2002;@Hecht2018;@Xiu2009].
 
-Global polynomials can achieve a better global accuracy
-for a wide class of regular functions
-(differentiable, smooth, analytic),
-and are amenable to further numerical computations.
-However, they are often misperceived as unstable,
-susceptible to Runge's phenomenon,
-prone to round-off errors,
-and subject to the curse of dimensionality [@Trefethen2011;@Trefethen2017].
+Despite their aforementioned advantages as global polynomials,
+global interpolating polynomials have a controversial reputation
[…]
