### 2025-04-07 minterpy-project/minterpy#10 @damar-wicaksono
- [x] Line 29: provide _a_ construction method --- Fixed with a reformulation of the sentence
- [x]  Line 33: better compared to what? --- Compared to local piecewise polynomials as mentioned in the opening paragraph of Section 2.
- [x] Line 38: what is appropriate in this context? --- Fixed by mentioning Chebyshev-Lobatto nodes
- [ ] Lines 53-54: in a practical sense, what are the advantages of this approach?
- [x] Figure 1: is the code for the tests in this figure available? --- Fixed by adding reference to the archived result in RODARE.
