### 2025-04-07 minterpy-project/minterpy#16 @damar-wicaksono
> I think it would be helpful if you could more clearly state the primary purpose of this approach? I read the following, but please correct or confirm:
>
> - the multi-index set paramaterization lends itself to "reducing curse of dimensionality" (by using p=2)
> - primary use case is approximation not interpolation, so use [non-equally spaced grid](https://minterpy-project.github.io/minterpy/latest/_images/e3e96bcda98441dac4d453d5bbb9ffef67fd8e4e81024b86034d8084feebd0f9.png) to address Runge phenomona
> - this representation (or perhaps transformation to Newton basis) allows easy implementation of closure under arithmetic and differentiation

Coming back to the points raised above:

- By setting $p = 2.0$, we can mitigate the curse of dimensionality, particularly in comparison to $p = \infty$ (tensor product construction), while maintaining the same approximation power (as referenced in the paper).
- Minterpy's primary use case is approximating functions via high-degree interpolating polynomials. Using non-equispaced grids, such as Chebyshev-Lobatto points, can help mitigate the Runge's phenomenon that occurs when high-degree interpolating polynomials are used with equispaced interpolation nodes.
- The transformation between Newton and Lagrange bases is crucial for ensuring closure under arithmetic and calculus operations. For example, adding two polynomials in the Newton basis involves computing the Lagrange coefficients (function values) at the unisolvent nodes for both polynomials, adding the coefficients together, and then transforming the resulting Lagrange coefficients back to Newton coefficients. Note that in the interpolating basis, coefficients cannot be simply added together unless the interpolation points of the interpolating polynomials are identical. Since Chebyshev-Lobatto nodes are not nested, different degrees typically have different nodes.

> Can you provide a brief/high-level summary statement of the references explaining the following:
>
> - How global polynomials are misperceived as unstable, susceptible to Runge's phenomone, etc. (lines 34-36)?
> - What about the grid reduces addresses Runge phenomona? (assuming that was a correct interpretation)
> - The meaning of curse of dimensionality in this context and how Euclidian degree reduces it

In relation to the points above, I've expanded the Statement of Need to have clearer argument regarding the misperceptions on the use of interpolating polynomials for function approximations.

Point 3 is additionally addressed in the Package Overview when the choice of $p$ is discussed.
