### 2025-04-07 #7702 @damar-wicaksono
Hi @sixpearls,

I've updated the paper to address your comments regarding the issues below. For further discussions, if any, perhaps you could raise them in the relevant Issues:

- [# [JOSS] Revise typos and grammatical mistakes in the paper (Reviewer 1)](https://github.com/minterpy-project/minterpy/issues/14#issue-2965969775)
- [# [JOSS] Provide additional clarification related to the statement of need (Reviewer 1)](https://github.com/minterpy-project/minterpy/issues/16#issue-2965979855)

Thanks a lot!
