### 2025-04-08 b3e32c7
Revise Statement of Need with clearer comparison with other packages
Following a remark by Reviever 2, the statement of need of the paper
has been updated to have a clearer comparison with other packages
and how Minterpy is different.

This commit should resolve Issue #10.
