+++ joss/paper.md
@@ -67,41 +67,38 @@ commonly associated with interpolation tasks.
 
 As a means of approximating functions, global polynomials---where a single
 polynomial is defined over the entire domain---offer several advantages.
-For sufficiently smooth functions, global polynomials can achieve high accuracy
-with a smaller number of data points (sampled over the entire domain) compared
+For sufficiently smooth functions, global polynomials can achieve
+high accuracy with a smaller number of data points (sampled over the entire domain) compared
 to local piecewise polynomials.
-Additionally, their relatively simple structure facilitates
-many common numerical operations.
+Additionally, their relatively simple structure facilitates many common numerical operations.
 These operations include differentiation, integration, addition, subtraction,
 and multiplication [@Trefethen2019].
 
 The Stone-Weierstrass theorem establishes that any continuous function
 on a bounded domain in multiple dimensions can be approximated uniformly
 to arbitrary precision by multivariate global polynomials [@Branges1959].
-However, the theorem does not specify a concrete method
-for constructing such approximating polynomials.
+However, the theorem does not specify a concrete method for constructing
+such approximating polynomials.
 Various techniques can be employed to build approximating polynomials,
 including Taylor expansions and least square approximations.
-Minterpy focuses on constructing approximating global polynomials
-using one of the earliest and most established methods:
-interpolation [@Goldstine1977].
+Minterpy focuses on constructing approximating global polynomials using
+one of the earliest and most established methods: interpolation [@Goldstine1977].
 
 Polynomial interpolation is based on the principle that, in one dimension,
-there exists a unique polynomial $Q_{f, n}$ of degree $n$
-that interpolates a function $f: \Omega \to \mathbb{R}$ in a bounded domain
+there exists a unique polynomial $Q_{f, n}$ of degree $n$ that interpolates
+a function $f: \Omega \to \mathbb{R}$ in a bounded domain
 $\Omega$ with $n + 1$ _distinct (unisolvent[^unisolvent]) interpolation nodes
-(or points)_ $P_n$ such that 
+(or points)_ $P_n$ such that
 $$
 Q_{f, n} (p_i) = f(p_i),\; \forall p_i \in P_n \subset \Omega,\; i = 0, \ldots, n.
 $$
 Polynomial interpolation has its roots in the works of Newton, Euler, Lagrange,
-and others [@Meijering2002],
-and its significance in mathematics and computing
-is well-established [@Cools2002;@Hecht2018;@Xiu2009].
+and others [@Meijering2002], and its significance in mathematics and computing
+is well-established [@Cools2002;@Hecht2018;@Xiu2009].  
 
 Despite their aforementioned advantages as global polynomials,
-global interpolating polynomials have a controversial reputation
-due to several common misconceptions [@Trefethen2011;@Trefethen2016;@Trefethen2017]:
+global interpolating polynomials have a controversial reputation due to
+several common misconceptions [@Trefethen2011;@Trefethen2016;@Trefethen2017]:  
 
 - They are often thought to be prone to Runge's phenomenon,
   whereby increasing the degree of interpolating polynomials worsens
@@ -125,45 +122,63 @@ Minterpy addresses these issues by:
   appropriate interpolation nodes (e.g., Chebyshev-Lobatto nodes)
   to help mitigate Runge's phenomenon[^equispaced];
 - Representing the interpolating polynomials in the Newton basis,
-  combined with Leja ordering of the interpolation nodes
+  combined with Leja ordering of the interpolation nodes  
   to ensure stable evaluation [@Reichel1990;@TalEzer1991;@Breuss2018];
 - Using a multi-index set to represent the multivariate polynomials,
   which can be tailored to mitigate the curse of dimensionality
   while preserving the approximation power of the interpolating polynomials
[…]
