### 2025-04-09 d81f69f
Revise README.md and CONTRIBUTING.md
Following remarks from the reviewers, broken links
in README.md has been fixed and the whole file
has been revised for clarity.

Additionally, CONTRIBUTING.md has also been revised
by fixing any broken and stale links
and any obsolete information.
When applicable, references are made to the online
documentation.

This commit should resolve Issue #11 and #15.
