+++ README.md
@@ -12,21 +12,25 @@
 | [`main`](https://github.com/minterpy-project/minterpy/tree/main) (stable) | [![Build](https://github.com/minterpy-project/minterpy/actions/workflows/build.yaml/badge.svg?branch=main)](https://github.com/minterpy-project/minterpy/actions/workflows/build.yaml?query=branch%3Amain+) [![codecov](https://codecov.io/gh/minterpy-project/minterpy/branch/main/graph/badge.svg?token=J8RCUGRKW3)](https://codecov.io/gh/minterpy-project/minterpy) [![Documentation Build and Deployment](https://github.com/minterpy-project/minterpy/actions/workflows/docs.yaml/badge.svg?branch=main)](https://minterpy-project.github.io/minterpy/stable/) |
 |  [`dev`](https://github.com/minterpy-project/minterpy/tree/dev) (latest)  | [![Build](https://github.com/minterpy-project/minterpy/actions/workflows/build.yaml/badge.svg?branch=dev)](https://github.com/minterpy-project/minterpy/actions/workflows/build.yaml?query=branch%3Adev) [![codecov](https://codecov.io/gh/minterpy-project/minterpy/graph/badge.svg?token=J8RCUGRKW3)](https://codecov.io/gh/minterpy-project/minterpy) [![Documentation Build and Deployment](https://github.com/minterpy-project/minterpy/actions/workflows/docs.yaml/badge.svg?branch=dev)](https://minterpy-project.github.io/minterpy/latest/)                 |
 
-Minterpy is an open-source Python package designed for constructing and manipulating multivariate interpolating polynomials
+Minterpy is an open-source Python package designed for constructing
+and manipulating multivariate interpolating polynomials
 with the goal of lifting the curse of dimensionality from interpolation tasks.
 
-Minterpy is being continuously extended and improved, with new functionalities added to address the bottlenecks involving
+Minterpy is being continuously extended and improved,
+with new functionalities added to address the bottlenecks involving
 interpolations in various computational tasks.
 
 ## Installation
 
-You can obtain the stable release of Minterpy directly from PyPI using `pip`:
+You can obtain the stable release of Minterpy directly
+from [PyPI](https://pypi.org/project/minterpy/) using `pip`:
 
 ```bash
 pip install minterpy
 ```
 
-Alternatively, you can also obtain the latest version of Minterpy from the GitHub repository:
+Alternatively, you can also obtain the latest version of Minterpy
+from the [GitHub repository](https://github.com/minterpy-project/minterpy):
 
 ```bash
 git clone https://github.com/minterpy-project/minterpy
@@ -38,137 +42,179 @@ Then from the source directory, you can install Minterpy:
 pip install [-e] .[all,dev,docs]
 ```
 
-where the flag `-e` means the package is directly linked
-into the python site-packages of your Python version.
+where the flag `-e` means the package is directly linked into
+the python site-packages of your Python version.
 The options `[all,dev,docs]` refer to the requirements defined
 in the `options.extras_require` section in `setup.cfg`.
 
-A best practice is to first create a virtual environment with
-the help of a tool like [mamba], [conda], [venv], [virtualenv] or [pyenv-virtualenv].
+A best practice is to first create a virtual environment with the help of
+a tool like [mamba], [conda], [venv], [virtualenv] or [pyenv-virtualenv].
[…]
