### 2025-04-09 #7702 @damar-wicaksono
@sixpearls, @kenohori,

Following your feedback, the [README.md](https://github.com/minterpy-project/minterpy/blob/dev-joss/README.md) and [CONTRIBUTING.md](https://github.com/minterpy-project/minterpy/blob/dev-joss/CONTRIBUTING.md) files have been revised, primarily by updating or removing broken and outdated links. Obsolete and inconsistent information has been removed, and references to the online documentation have been added where applicable. The issues related to the feedback are:

- [[JOSS] Fix issues in README.md (Reviewer 1) #15](https://github.com/minterpy-project/minterpy/issues/15)
- [[JOSS] Revise README.md according to JOSS Review (Reviewer 2) #11](https://github.com/minterpy-project/minterpy/issues/11)

Please note that these modifications are still in `dev-joss` branch; any modifications made during to this review process will later be merged to `dev` then to `main` (for release).

Thank you!
