### 2025-04-09 minterpy-project/minterpy#11 @damar-wicaksono
The [README.md](https://github.com/minterpy-project/minterpy/blob/dev-joss/README.md) and [CONTRIBUTING.md](https://github.com/minterpy-project/minterpy/blob/dev-joss/CONTRIBUTING.md) files have been revised, primarily by updating or removing broken and outdated links. Obsolete and inconsistent information has been removed, and references to the online documentation have been added where applicable.

Note that these modifications are still in `dev-joss` branch; any modifications made during to this review process will later be merged to `dev` then to `main` (for release).
