### 2025-04-11 97b09ee
Update package version in the paper
