### 2025-04-11 minterpy-project/minterpy#13 @damar-wicaksono
The details on the numerical experiments, their results, and notes on reproduction are now available in a dedicated repository https://github.com/minterpy-project/minterpy-joss-data.

The experiments will remain in this dedicated repository and not in the main repository of Minterpy due to external dependency.

Once the review process is concluded, a new archive will be created in [RODARE](https://rodare.hzdr.de/record/3379).
