### 2025-04-11 #7702 @damar-wicaksono
@sixpearls,

Regarding your point on reproducibility: I've updated the README for the numerical experiments used to generate the convergence plot in the figure. The updated documentation and related files are now hosted in a public repository. You can find the details and leave any comments on the following issue:

- [[JOSS] Add additional details to the archived numerical experiment (Reviewer 1) #13](https://github.com/minterpy-project/minterpy/issues/13)

Once this issue is resolved, I’ll create a new archival entry on RODARE, as required by our institute. The paper will include a reference to this archived version.

Thank you!

PS: The plot has been updated to use the latest version of Chaospy (v4.3.18) and ndsplines (v0.2.0.post0), just to be sure;  there's no apparent difference.
PPS: With this, I’ve attempted to address all issues raised in the previous review round. I’ll wait for your further feedback.
