### 2025-04-14 a1177e1
Update Quickstart in README and fix typo
