### 2025-04-20 #7702 @sixpearls
Thanks for making these changes. I only have two blocking issues for the publication:

- Fix the math rendering error on line 101
- Provide some mention of the tradeoff of run-time for the increased accuracy and convenience for certain mathematical operations. The global polynomial requires many more floating point multiplications and additions compared to piecewise polynomials (e.g., the compact support of B-splines).

Overall I found this work very interesting. It is beyond the scope of requirement for this JOSS article, but it might be interesting to have a slightly more detailed (but easy-to-follow for non-specialists) explanation of the various tradeoffs of interpolation approaches. I found this additional comment

> The significance of this capability... many numerical methods can be boiled down to these fundamental operations on functions
(line 95-96)

to be very compelling. I could see why someone would prefer this approach over piecewise polynomials in certain conditions.
Well done!
