### 2025-04-22 minterpy-project/minterpy#17 @damar-wicaksono
[JOSS] Provide a short comment on the tradeoff between accuracy and computational cost for global polynomials and local piecewise polynomials (Reviewer 1)
Following a suggestion from [Reviewer 1](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2817366676):

> Provide some mention of the tradeoff of run-time for the increased accuracy and convenience for certain mathematical operations. The global polynomial requires many more floating point multiplications and additions compared to piecewise polynomials (e.g., the compact support of B-splines).

The manuscript will be updated to include a brief discussion of this trade-off.

The reviewer made a great point: global polynomial interpolants are indeed more computationally expensive to evaluate than local piecewise polynomials such as B-splines, primarily due to their lack of compact support—evaluating a global polynomial typically involves all terms in the expansion. While the current JOSS submission does not make any explicit claims about the performance of Minterpy, it is still worthwhile to acknowledge this consideration for completeness.
