### 2025-04-22 d7ffab9
Replace the hypercube notation in the paper.
