### 2025-04-22 20defc2
Add commentary on the tradeoff between accuracy and cost
A paragraph has been expanded to include a commentary
on the trade-off between the accuracy of global
polynomials and the cost of their evaluation, especially
in comparison to B-splines and local piecewise polynomials.

This commit should resolve Issue @17.
