### 2025-04-22 6d7b007
Add reference to the archived numerical experiment
