+++ joss/paper.bib
@@ -334,4 +334,12 @@
   doi     = {10.1137/0912034},
 }
 
+@Misc{Wicaksono2025,
+  author = {Wicaksono, Damar Canggih and Hernandez Acosta, Uwe and Thekke Veettil, Sachin Krishnan and Kissinger, Jannik and Hecht, Michael},
+  title  = {{Data to "Minterpy: Multivariate polynomial interpolation in Python"}},
+  year   = {2025},
+  doi    = {10.14278/rodare.3379},
+  url    = {https://doi.org/10.14278/rodare.3379},
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
