### 2025-04-22 6961460
Simplify the paper
