+++ joss/paper.bib
@@ -64,19 +64,6 @@
   volume  = {5},
 }
 
-@InBook{Hecht2018,
-  author    = {Hecht, Michael and Sbalzarini, Ivo F.},
-  pages     = {53--75},
-  publisher = {Springer International Publishing},
-  title     = {Fast interpolation and fourier transform in high-dimensional spaces},
-  year      = {2018},
-  isbn      = {9783030011772},
-  month     = nov,
-  booktitle = {Intelligent Computing},
-  doi       = {10.1007/978-3-030-01177-2_5},
-  issn      = {2194-5365},
-}
-
 @Book{Trefethen2019,
   author    = {Trefethen, L. N.},
   publisher = {Society for Industrial and Applied Mathematics (SIAM)},
