+++ joss/paper.md
@@ -68,9 +68,10 @@ commonly associated with interpolation tasks.
 As a means of approximating functions, global polynomials---where a single
 polynomial is defined over the entire domain---offer several advantages.
 For sufficiently smooth functions, global polynomials can achieve
-high accuracy with a smaller number of data points (sampled over the entire domain) compared
-to local piecewise polynomials.
-Additionally, their relatively simple structure facilitates many common numerical operations.
+high accuracy with a smaller number of data points (sampled over the entire domain)
+compared to local piecewise polynomials.
+Additionally, their relatively simple structure facilitates
+many common numerical operations.
 These operations include differentiation, integration, addition, subtraction,
 and multiplication [@Trefethen2019].
 
@@ -80,7 +81,7 @@ to arbitrary precision by multivariate global polynomials [@Branges1959].
 However, the theorem does not specify a concrete method for constructing
 such approximating polynomials.
 Various techniques can be employed to build approximating polynomials,
-including Taylor expansions and least square approximations.
+such as least square approximations.
 Minterpy focuses on constructing approximating global polynomials using
 one of the earliest and most established methods: interpolation [@Goldstine1977].
 
@@ -89,16 +90,17 @@ there exists a unique polynomial $Q_{f, n}$ of degree $n$ that interpolates
 a function $f: \Omega \to \mathbb{R}$ in a bounded domain
 $\Omega$ with $n + 1$ _distinct (unisolvent[^unisolvent]) interpolation nodes
 (or points)_ $P_n$ such that
-$$
+\begin{equation}
+\label{eq:interpolation-condition}
 Q_{f, n} (p_i) = f(p_i),\; \forall p_i \in P_n \subset \Omega,\; i = 0, \ldots, n.
-$$
+\end{equation}
 Polynomial interpolation has its roots in the works of Newton, Euler, Lagrange,
 and others [@Meijering2002], and its significance in mathematics and computing
-is well-established [@Cools2002;@Hecht2018;@Xiu2009].  
+is well-established [@Cools2002;@Xiu2009].  
 
 Despite their aforementioned advantages as global polynomials,
 global interpolating polynomials have a controversial reputation due to
-several common misconceptions [@Trefethen2011;@Trefethen2016;@Trefethen2017]:  
+several misconceptions [@Trefethen2011;@Trefethen2016;@Trefethen2017]:  
 
 - They are often thought to be prone to Runge's phenomenon,
   whereby increasing the degree of interpolating polynomials worsens
@@ -147,7 +149,7 @@ equadratures [@Seshadri2017], PyGPC [@Weise2020], PyThia [@Hegemann2023],
 and UncertainSci [@Tate2023], provide polynomial-based function approximations,
 primarily for uncertainty quantification (UQ) using generalized polynomial
 chaos expansion [@Xiu2002].
-These tools naturally frame problems as UQ tasks, where inputs are modeled
+These tools frame problems as UQ tasks, where inputs are modeled
 probabilistically.
 With few exceptions (notably Chaospy), the resulting polynomials are primarily
 used for function approximations, accompanied by additional post-processing
@@ -155,7 +157,7 @@ utilities tailored to UQ tasks (e.g., uncertainty propagation,
 sensitivity analysis).
 In contrast, Minterpy offers a simpler, UQ-free approach to function approximation
 using interpolating polynomials, with fewer barriers to entry, and includes
-additional standard mathematical operations on the polynomials.
+several mathematical operations on the polynomials.
 
[…]
