### 2025-04-23 63a88d9
Fix issue with the Quickstart guide
The Quickstart guide has been revised to
fix typos and grammatical mistakes.
Furthermore, several sentences have been
reformulated for clarity.

This commit should resolve Issue #21.
