+++ pr-preview/29/_sources/getting-started/polynomial-bases-and-transformations.ipynb.txt
@@ -1,0 +1,2687 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "60f73f47",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": "# Polynomial Bases and Transformations"
+  },
+  {
+   "cell_type": "markdown",
+   "id": "9250bb6bcb0da956",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "The Minterpy polynomials you've encountered so far have been represented in the Lagrange and Newton bases.\n",
+    "Minterpy supports a range of polynomial bases, each with a consistent interface, meaning that polynomials in different bases behave and can be manipulated in similar ways.\n",
+    "\n",
+    "In this in-depth tutorial, you'll explore:\n",
+    "\n",
+    "- various polynomial bases supported by Minterpy\n",
+    "- transformation between bases\n",
+    "\n",
+    "For simplicity, this tutorial uses examples in one and two dimensions, but the principles apply to polynomials of higher dimensions as well.\n",
+    "\n",
+    "Before diving in, make sure to import the necessary packages:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "id": "df692690-ece7-408f-bf6a-38fec281b28f",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "outputs": [],
+   "source": [
+    "import minterpy as mp\n",
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt\n",
+    "\n",
+    "import warnings\n",
+    "warnings.filterwarnings('ignore')"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "5460c3b2-003e-48c5-a089-d0360a000ea9",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "## Polynomials in Minterpy\n",
+    "\n",
+    "Polynomials in Minterpy are multidimensional polynomials defined on $[-1, 1]^m$ where $m$ is the spatial dimension:\n",
+    "\n",
+    "$$\n",
+    "Q(\\boldsymbol{x}) = \\sum_{\\boldsymbol{\\alpha} \\in A} c_{\\cdot, \\boldsymbol{\\alpha}} \\Psi_{\\cdot, \\boldsymbol{\\alpha}} (\\boldsymbol{x}),\n",
+    "$$\n",
+    "\n",
+    "where:\n",
+    "\n",
+    "- $A$ is the multi-index set of the exponents, i.e., $\\boldsymbol{\\alpha} \\in A \\subseteq \\mathbb{N}^m$ which generalizes the notion of polynomial degree to multiple dimensions.\n",
+    "- $\\Psi_{\\cdot, \\boldsymbol{\\alpha}}$ is the multidimensional basis polynomial associated with the multi-index element $\\boldsymbol{\\alpha}$, the placeholder $\\cdot$ indicates that the basis polynomials depend on the chosen basis.\n",
+    "- $c_{\\cdot, \\boldsymbol{\\alpha}}$ is the coefficient that corresponds to the multidimensional basis polynomial $\\Psi_{\\cdot, \\boldsymbol{\\alpha}}$.\n",
[…]
