+++ pr-preview/29/_sources/api/core/index.rst.txt
@@ -1,0 +1,14 @@
+=============
+minterpy.core
+=============
+
+.. automodule:: minterpy.core
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   Multi-Index Set <multi_index>
+   Grid <grid>
+   Abstract Base Classes <ABC/index>
+   Multi-Index Tree <tree>
