+++ pr-preview/29/_sources/api/internal/jit_compiled/index.rst.txt
@@ -1,0 +1,16 @@
+=====================
+minterpy.jit_compiled
+=====================
+
+.. automodule:: minterpy.jit_compiled
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   Common <common>
+   Multi-index Exponents <multi-index>
+   Newton: Differentiation <newton-diff>
+   Newton: Evaluation <newton-eval>
+   Canonical <canonical>
+   Polynomial transformations <transformations>
