+++ pr-preview/29/_sources/api/polynomials/lagrange.rst.txt
@@ -1,0 +1,24 @@
+===================
+lagrange_polynomial
+===================
+
+.. currentmodule:: minterpy.polynomials.lagrange_polynomial
+
+.. automodule:: minterpy.polynomials.lagrange_polynomial
+
+.. autoclass:: LagrangePolynomial
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.polynomials.lagrange_polynomial.LagrangePolynomial
+      :properties:
+
+   .. rubric:: Methods
+
+   .. classautosummary:: minterpy.polynomials.lagrange_polynomial.LagrangePolynomial
+      :methods:

[…]
