### 2025-04-24 ebaa73e
Use updated upload artifact GitHub action
`actions/upload-artifact@v3` has been deprecated
and causes the CI/CD pipeline to break.
The version used has been updated to v4.

Due to some changes in v4, each uploaded artifacts
are now explicitly named according to their
steps in the workflows.

This commit should resolve Issue #30.
