+++ pr-preview/31/_sources/getting-started/arithmetic-operations-with-polynomials.ipynb.txt
@@ -1,0 +1,1162 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "60f73f47",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "# Basic Arithmetic Operations with Polynomials"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "9250bb6bcb0da956",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "Once you have constructed a polynomial in Minterpy, you can perform various operations on it, depending on the basis in which the polynomial is represented.\n",
+    "For instance, as demonstrated in the previous tutorials, you can evaluate a Minterpy polynomial (in the Newton basis) at a set of query points.\n",
+    "\n",
+    "This tutorial will show you how to extend beyond evaluation by performing arithmetic operations with Minterpy polynomials.\n",
+    "Specifically, you'll learn that Minterpy polynomials can be:\n",
+    "\n",
+    "- added or subtracted from other polynomials\n",
+    "- multiplied by other polynomials or real scalar numbers\n",
+    "- raised to a non-negative integer power\n",
+    "\n",
+    "These operations result in another Minterpy polynomial (in the same basis).\n",
+    "In other words, Minterpy polynomials are _closed_ under the following arithmetic operations:\n",
+    "\n",
+    "- polynomial-polynomial multiplication\n",
+    "- polynomial-scalar multiplication\n",
+    "- polynomial-polynomial addition and subtraction\n",
+    "- polynomial-scalar addition and subtraction\n",
+    "- exponentiation by a non-negative integer\n",
+    "\n",
+    "For clarity, this in-depth tutorial uses one-dimensional polynomials as examples.\n",
+    "However, the principles and behaviors described are consistent across all dimensions.\n",
+    "\n",
+    "Before you start, make sure to import the necessary packages to follow along with this guide."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "id": "aa387b47-58fc-453c-b738-662bc24ff0fc",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "outputs": [],
+   "source": [
+    "import minterpy as mp\n",
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "52ea8534-a3d0-4b98-807b-25777b4594f0",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "## Motivating functions\n",
+    "\n",
+    "Consider the following two one-dimensional functions.\n",
+    "\n",
+    "The first is the sine function:\n",
+    "\n",
+    "$$\n",
[…]
