+++ pr-preview/29/_sources/how-to/grid/grid-create-from-value-set.ipynb.txt
@@ -1,259 +1,0 @@
-{
- "cells": [
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": "# Create a Grid from an Array of Generating Values",
-   "id": "30fda6830c7191f2"
-  },
-  {
-   "metadata": {},
-   "cell_type": "code",
-   "source": [
-    "import numpy as np\n",
-    "import minterpy as mp\n",
-    "import matplotlib.pyplot as plt"
-   ],
-   "id": "284f17796016b5b",
-   "outputs": [],
-   "execution_count": null
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "Minterpy interpolating polynomials (i.e., in the Lagrange or Newton basis)\n",
-    "lives on a grid that holds the so-called _unisolvent nodes_.\n",
-    "An interpolating multi-dimensional polynomial of a given multi-index set of\n",
-    "exponents is uniquely\n",
-    "determined on the corresponding unisolvent nodes.\n",
-    "\n",
-    "An instance of `Grid` can be constructed via different constructors:\n",
-    "\n",
-    "- {py:meth}`.Grid.from_value_set`: create an instance based on an array of generating values (this page)\n",
-    "- {py:meth}`.Grid.from_function`: create an instance with a generating function (see the {doc}`example <grid-create-from-function>`)\n",
-    "- {py:meth}`.Grid.from_points`: create an instance with an array of generating points (see the {doc}`example <grid-create-from-points>`)\n",
-    "- {py:meth}`.Grid.from_degree`: create an instance with a complete multi-index set (see the {doc}`example <grid-create-from-degree>`)\n",
-    "- {py:class}`.Grid`: the default constructor (see the {doc}`example <grid-create-default-constructor>`)\n",
-    "\n",
-    "This guide provides an example on how to construct a `Grid` instance based\n",
-    "on a given array of generating values (i.e., one-dimensional interpolation\n",
-    "points) using the `from_value_set()` method."
-   ],
-   "id": "b6943bbbd9caee2b"
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "## About generating values\n",
-    "\n",
-    "_Generating values_ in Minterpy convention, are one-dimensional interpolation\n",
-    "points. When used to define a grid to support multi-dimensional polynomials,\n",
-    "the generating values will be converted to an array of generating points\n",
-    "via tiling procedure.\n",
-    " \n",
-    "Valid generating values to create an instance of `Grid` are given as\n",
-    "a one-dimensional array of floats that must additionally satisfy the following\n",
-    "conditions:\n",
-    "\n",
-    "- the length must be $n + 1$ where $n$ is the maximum polynomial degree\n",
-    "  across all dimensions that the grid must support.\n",
-    "- the values of the array must be unique, that is, it contains no repeats.\n",
-    "- the length must be equal to or larger than the maximum exponents\n",
-    "  of the grid's multi-index set in any dimension.\n",
-    "\n",
-    "\n",
-    "```{note}\n",
-    "Generating values may also be provided as a two-dimensional array as long\n",
-    "as the array only has either one colum or one row.\n",
-    "```"
-   ],
-   "id": "799ba4eb511bf60e"
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
[…]
