+++ pr-preview/32/_sources/api/transformation-operators/operator-barycentric/operator-barycentric-factorized.rst.txt
@@ -1,0 +1,22 @@
+=============================
+BarycentricFactorisedOperator
+=============================
+
+.. currentmodule:: minterpy.schemes.barycentric.operators
+
+.. autoclass:: BarycentricFactorisedOperator
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.schemes.barycentric.operators.BarycentricFactorisedOperator
+      :properties:
+
+   .. rubric:: Methods
+
+   .. classautosummary:: minterpy.schemes.barycentric.operators.BarycentricFactorisedOperator
+      :methods:
