+++ pr-preview/32/_sources/getting-started/1d-polynomial-interpolation.ipynb.txt
@@ -1,0 +1,1242 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "60f73f47",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "# One-Dimensional Polynomial Interpolation"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "9250bb6bcb0da956",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "Welcome to the first in-depth tutorial of Minterpy!\n",
+    "\n",
+    "Minterpy is an open source Python package for carrying out numerical tasks with multivariate (multi-dimensional) polynomials.\n",
+    "You can do a wide range of arithmetic and calculus operations with Minterpy polynomials.\n",
+    "\n",
+    "_But how to create such polynomials in the first place?_\n",
+    "\n",
+    "This tutorial demonstrates how you can create such a polynomial from scratch in the most typical use case: interpolating a function."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "7dbbdff064d29556",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "## Import Minterpy\n",
+    "\n",
+    "After installing {doc}`installing Minterpy </getting-started/index>`, you can import it as follows:"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "id": "91e743767ddf5d26",
+   "metadata": {
+    "ExecuteTime": {
+     "end_time": "2024-08-23T11:06:29.487413Z",
+     "start_time": "2024-08-23T11:06:26.872834Z"
+    },
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "outputs": [],
+   "source": [
+    "import minterpy as mp"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "f55c64e669a92386",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "The shorthand `mp` is the convention we, the developer team, like to use; it's short and recognizable (just like `np` for NumPy and `pd` for pandas)."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "e773ec89-4a90-47e4-a034-81cfd61150f7",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "You will also need to import NumPy and Matplotlib to follow along with this guide."
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": null,
+   "id": "26f8a4b7-8487-40d9-8c7a-d54ce3bc2224",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "outputs": [],
+   "source": [
+    "import numpy as np\n",
+    "import matplotlib.pyplot as plt"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "c7ac4bf8-40f0-443e-8eb6-cfe70cbd7490",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
[…]
