+++ pr-preview/32/_sources/how-to/multi-index-set/index.rst.txt
@@ -1,0 +1,20 @@
+################
+Multi-Index Sets
+################
+
+This page lists all the available How-To Guides related to instances of
+the :py:class:`.MultiIndexSet` class (i.e., the multi-index sets).
+
+.. toctree::
+   :maxdepth: 1
+
+   Create a Multi-Index Set <multi-index-set-create>
+   Add Exponents to a Multi-Index Set <multi-index-set-add>
+   Check for Equality in Value <multi-index-set-equality>
+   Check for Membership (Containment) <multi-index-set-membership>
+   Compare Multi-Index Sets <multi-index-set-comparison>
+   Make a Multi-Index Set Complete <multi-index-set-complete>
+   Make a Multi-Index Set Downward-Closed <multi-index-set-downward-closed>
+   Create the Union of Multi-Index Sets <multi-index-set-union>
+   Multiply Multi-Index Sets <multi-index-set-multiply>
+   Create an Empty Multi-Index Set <multi-index-set-empty>
