+++ pr-preview/33/_sources/fundamentals/interpolation-problem.rst.txt
@@ -1,0 +1,177 @@
+=========================================
+Multidimensional Polynomial Interpolation
+=========================================
+
+Polynomial interpolation dates back to Newton, Lagrange,
+and others\ :footcite:`meijering2002`,
+and its fundamental importance in both mathematics and computing is widely
+recognized.
+
+Interpolation is rooted in the principle that, in one dimension,
+there exists exactly one polynomial :math:`Q_{f, A}` of degree
+:math:`n \in \mathbb{N}` that *interpolates* a function
+:math:`f : \mathbb{R} \longrightarrow \mathbb{R}` on :math:`n+1` distinct
+*interpolation nodes* :math:`P_A = \{p_i\}_{i \in A} \subseteq \mathbb{R}`
+where :math:`A=\{0, \ldots, n\}`.
+Specifically, this means
+
+.. math::
+
+  Q_{f,A}(p_i) = f(p_i)\,, \quad \text{for all} \quad  p_i \in P_A \,, i \in A\,.
+
+This makes interpolation fundamentally different from approximation,
+see :numref:`1d_interpol`.
+
+.. _1d_interpol:
+.. figure:: ./images/1D_interpol.png
+   :align: center
+
+   Interpolation and approximation in one dimension.
+   While the interpolant :math:`Q_{f, A}` must match the function :math:`f`
+   at the interpolation nodes :math:`p_0, \ldots, p_n`,
+   an approximation :math:`Q^*_n` is not required to coincide with :math:`f`
+   at all.
+
+Weierstrass Approximation Theorem
+=================================
+
+The famous *Weierstrass Approximation Theorem*\ :footcite:`weierstrass1885`
+asserts that any continuous function :math:`f : \Omega \longrightarrow \mathbb{R}`,
+defined on a compact domain, such as :math:`\Omega = [-1,1]^m`,
+can be uniformly approximated by polynomials\ :footcite:`debranges1959`.
+However, the theorem does not require the polynomials to coincide with
+:math:`f` at any given points.
+In fact, it is possible to have a sequence of multidimensional polynomials
+:math:`Q_n^*`, :math:`n \in \mathbb{N}`, where :math:`Q_n^*(x) \neq f(x)`
+for all :math:`x \in \Omega`, but still achieve uniform approximation,
+that is
+
+.. math::
+
+  Q_{n}^* \xrightarrow[n \rightarrow \infty]{} f \quad \text{uniformly on} \quad \Omega\,.
+
+There are several constructive proofs of the Weierstrass approximation theorem,
+including the well-known version by Serge Bernstein\ :footcite:`bernstein1912`.
+The resulting *Bernstein approximation scheme* is universal and has been shown
+to reach the optimal (inverse-linear)
+approximation rate for the absolute value function :math:`f(x) = |x|`
+\ :footcite:`bernstein1914`.
+However, despite its generality, the Bernstein scheme achieves only slow
+convergence rates for analytic functions, leading to high computational costs
+in practical applications.
+
+As a result, extensive research has focused on extending one-dimensional
+**Newton** or **Lagrange interpolation schemes** to multiple dimensions (mD)
+while preserving their computational efficiency.
+Any method addressing this challenge must avoid Runge's
+phenomenon\ :footcite:`runge1901` (overfitting) by ensuring uniform
+approximation of the interpolation target
+:math:`f : \Omega \longrightarrow \mathbb{R}`.
[…]
