+++ pr-preview/33/_sources/api/internal/utils/poly-chebyshev.rst.txt
@@ -1,6 +1,0 @@
-=====================
-polynomials.chebyshev
-=====================
-
-.. automodule:: minterpy.utils.polynomials.chebyshev
-   :members:
