+++ pr-preview/33/_sources/how-to/multi-index-set/multi-index-set-complete.ipynb.txt
@@ -1,304 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "5741d863-bc8c-4011-8d63-712c89ef9dd8",
-   "metadata": {},
-   "source": [
-    "# Make a Multi-Index Set Complete"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "c94d29f1-8250-4231-af69-440ef9df77df",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "79490875-cdde-497a-a5a8-c20588b5ece9",
-   "metadata": {},
-   "source": [
-    "A multi-index set $A$ is said to be _complete_ if, for a given polynomial of degree $n \\in \\mathbb{N}$ and with respect to a given $p$ of an $l_p$-norm (i.e., the $l_p$-degree), it contains all the exponents such that $\\lVert \\boldsymbol{\\alpha} \\rVert_p = (\\alpha_1^p + \\ldots + \\alpha_m^p)^{\\frac{1}{p}} \\leq n$ for all $\\boldsymbol{\\alpha} \\in A$.\n",
-    "\n",
-    "This guide shows how to check a given `MultiIndexSet` instance for completeness and make the corresponding set of exponents complete."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "d6f277a3-1a78-496f-9154-a092dbded327",
-   "metadata": {},
-   "source": [
-    "## Motivating example\n",
-    "\n",
-    "Consider a (incomplete) three-dimensional multi-index set with a single element and $l_p$-degree of $2.0$ as follows:\n",
-    "\n",
-    "$$\n",
-    "A = \\{ (0, 0, 2) \\}\n",
-    "$$"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "63a9fc78-d3e7-463a-9b11-cedf391ee827",
-   "metadata": {},
-   "source": [
-    "This multi-index set can be created in Minterpy as follows:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "3bff2feb-3a96-4f93-94b7-4ab900c8e4dd",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "mi = mp.MultiIndexSet(np.array([[0, 0, 2]]), lp_degree=2.0)"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "9a115ec2-1bcc-4dd5-9dc4-c1b5af2c1fdd",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "print(mi)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "5d0a5ad1-4a03-4829-9902-1a3918c7305b",
-   "metadata": {},
-   "source": [
-    "This set has a polynomial degree of:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "93096b3b-17c6-4827-a15e-bc918b960602",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "mi.poly_degree"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "33ce61cd-8c9b-434d-b70a-6a52509d2392",
-   "metadata": {},
-   "source": [
-    "```{note}\n",
[…]
