+++ latest/_sources/getting-started/md-polynomial-interpolation.ipynb.txt
@@ -10,7 +10,9 @@
     },
     "tags": []
    },
-   "source": "# Multidimensional Polynomial Interpolation"
+   "source": [
+    "# Multivariate Polynomial Interpolation"
+   ]
   },
   {
    "cell_type": "markdown",
@@ -25,7 +27,7 @@
    "source": [
     "In the {doc}`previous guide </getting-started/1d-polynomial-interpolation>`,\n",
     "you learned the basics of polynomial interpolation in Minterpy for approximating one-dimensional functions.\n",
-    "As the name \"Minterpy\" suggests, the package also supports constructing multivariate (multidimensional) function.\n",
+    "As the name \"Minterpy\" suggests, the package also supports constructing multivariate polynomials to approximate multidimensional functions.\n",
     "\n",
     "In this in-depth tutorial, you will learn how to create an interpolating polynomial that approximates a two-dimensional function.\n",
     "We use a two-dimensional example for ease of visualization, but the main steps you'll learn are applicable to polynomials of higher dimensions as well."
@@ -246,7 +248,7 @@
     "4. Create a polynomial in the Lagrange basis\n",
     "5. Transformation from the Lagrange basis to another basis, preferrably the Newton basis, for further manipulation.\n",
     "\n",
-    "As before, we will go through these steps one at a time, highlighting any differences that arise compared to the one dimensional case."
+    "As before, we will go through these steps one at a time, highlighting any differences that arise compared to the one-dimensional case."
    ]
   },
   {
@@ -264,7 +266,7 @@
     "\n",
     "As demonstrated in the previous tutorial, multi-index sets in Minterpy are represented by the {py:class}`MultiIndexSet <.core.multi_index.MultiIndexSet>` class.\n",
     "You can use the class method {py:meth}`from_degree() <.MultiIndexSet.from_degree>` to create a complete multi-index set for a given spatial dimension (first argument), polynomial degree and polynomial degree (second argument).\n",
-    "For multidimensional polynomials, you also need to specify the $l_p$-degree ($> 0.0$) of the multi-index set (third argument).\n",
+    "For multivariate polynomials, you also need to specify the $l_p$-degree ($> 0.0$) of the multi-index set (third argument).\n",
     "\n",
     "A complete multi-index set with spatial dimension $m$, polynomial degree $n$, and $l_p$-degree $p$ denoted by $A_{m, n, p}$ includes all multi-indices $\\boldsymbol{\\alpha}$ that satisfy the condition $\\lVert \\boldsymbol{\\alpha} \\rVert_p \\leq n$, specifically\n",
     "\n",
@@ -305,7 +307,9 @@
     },
     "tags": []
    },
-   "source": "The set contains the exponents of a multidimensional polynomial. Each column corresponds to the exponents per spatial dimension, while each row indicates the multi-index elements."
+   "source": [
+    "The set contains the exponents of a multidimensional polynomial. Each column corresponds to the exponents per spatial dimension, while each row indicates the multi-index elements."
+   ]
   },
   {
    "cell_type": "markdown",
@@ -348,7 +352,7 @@
     "tags": []
    },
    "source": [
-    "Several common choices for $l_p$-degree are $1.0$ (_total-degree set_), $2.0$ (_euclidian-degree set_), and $\\infty$ (_tensorial set_).\n",
[…]
