+++ pr-preview/34/_sources/how-to/grid/grid-call.ipynb.txt
@@ -1,0 +1,336 @@
+{
+ "cells": [
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": "# Evaluate a Function on a Grid",
+   "id": "5223970d3d19e2f7"
+  },
+  {
+   "metadata": {},
+   "cell_type": "code",
+   "source": [
+    "import minterpy as mp\n",
+    "import numpy as np"
+   ],
+   "id": "d32b739de18f3508",
+   "outputs": [],
+   "execution_count": null
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": [
+    "Calling an instance of `Grid` with a function or a `Callable` evaluates\n",
+    "the given function on the unisolvent nodes and returns the corresponding\n",
+    "function values. In the context of polynomial interpolation, these function\n",
+    "values are the coefficients of a polynomial in the Lagrange basis.\n",
+    "\n",
+    "This guide demonstrates how to call an instance of `Grid` on a function."
+   ],
+   "id": "aeee3be84ec26284"
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": [
+    "## Example: Function with one-dimensional output\n",
+    "\n",
+    "Consider the following problem:\n",
+    "\n",
+    "Compute the formula\n",
+    "\n",
+    "$$\n",
+    "f(\\boldsymbol{x}) = \\sum_{i = 1}^{3} x_i^2, \\boldsymbol{x} \\in [-1, 1]^3 \n",
+    "$$\n",
+    "\n",
+    "on the interpolation grid that corresponds to a complete multi-index set of\n",
+    "polynomial degree $3$ with respect to the $l_p$-degree $2.0$. "
+   ],
+   "id": "9882a42c341e5e4e"
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": [
+    "### Function to evaluate\n",
+    "\n",
+    "A valid function or callable that can be called with a `Grid` instance\n",
+    "must satisfy the following:\n",
+    "\n",
+    "- it must accept as the first argument a two-dimensional array where\n",
+    "  each row corresponds to different evaluation points and each column\n",
+    "  corresponds to different spatial dimension.\n",
+    "- it must return an array with the same length as the input.\n",
+    "\n",
+    "The function or callable may accept additional arguments either positional\n",
+    "or keyword.\n",
+    "\n",
+    "The function as required above can be defined as follows:"
+   ],
+   "id": "2b8c0abb1154afb9"
+  },
+  {
+   "metadata": {},
+   "cell_type": "code",
+   "source": [
+    "def fun_one_dim(xx: np.ndarray) -> np.ndarray:\n",
+    "    \"\"\"Compute the sum of squared.\"\"\"\n",
+    "    return np.sum(xx**2, axis=1)"
+   ],
+   "id": "2f23b787c2fe8c22",
+   "outputs": [],
+   "execution_count": null
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": [
+    "### Grid\n",
+    "\n",
+    "The interpolation grid that corresponds to the complete multi-index set\n",
+    "can be created using `from_degree()` factory method:"
+   ],
+   "id": "b7887f75656036e1"
+  },
+  {
+   "metadata": {},
+   "cell_type": "code",
+   "source": [
+    "spatial_dimension = 3\n",
+    "poly_degree = 3\n",
+    "lp_degree = 2.0\n",
[…]
