+++ pr-preview/34/_sources/getting-started/md-polynomial-interpolation.ipynb.txt
@@ -1,1272 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "4f44cdf37aadcff5",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "# Multivariate Polynomial Interpolation"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "c26c295879e98b28",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "In the {doc}`previous guide </getting-started/1d-polynomial-interpolation>`,\n",
-    "you learned the basics of polynomial interpolation in Minterpy for approximating one-dimensional functions.\n",
-    "As the name \"Minterpy\" suggests, the package also supports constructing multivariate polynomials to approximate multidimensional functions.\n",
-    "\n",
-    "In this in-depth tutorial, you will learn how to create an interpolating polynomial that approximates a two-dimensional function.\n",
-    "We use a two-dimensional example for ease of visualization, but the main steps you'll learn are applicable to polynomials of higher dimensions as well."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "8fd381f9-68ea-4dd3-a055-85c8cadcd170",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "Before we begin, let's import the necessary packages to follow along with this guide."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "1562e776-f727-4ab0-968e-8f798ec9729b",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np\n",
-    "import matplotlib.pyplot as plt"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "ea4cacd9-7afd-46af-bf13-582014249a75",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "## Motivating function\n",
-    "\n",
-    "Consider the following two-dimensional function:\n",
-    "\n",
-    "$$\n",
-    "\\begin{align}\n",
-    "\t\\mathcal{M}(\\boldsymbol{x}) = & 0.75 \\exp{\\left( -0.25 \\left( (x_1 - 2)^2 + (x_2 - 2)^2 \\right) \\right) } \\\\\n",
-    "                                  & + 0.75 \\exp{\\left( -1.00 \\left( \\frac{(x_1 + 1)^2}{49} + \\frac{(x_2 + 1)^2}{10} \\right) \\right)} \\\\\n",
-    "\t\t\t\t\t\t\t\t  & + 0.50 \\exp{\\left( -0.25 \\left( (x_1 - 7)^2 + (x_2 - 3)^2 \\right) \\right)} \\\\\n",
-    "\t\t\t\t\t\t\t\t  & - 0.20 \\exp{\\left( -1.00 \\left( (x_1 - 4)^2 + (x_2 - 7)^2 \\right) \\right)} \\\\\n",
-    "\\end{align}\n",
-    "$$\n",
-    "\n",
[…]
