+++ pr-preview/34/_sources/how-to/multi-index-set/multi-index-set-equality.ipynb.txt
@@ -1,274 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "f37c96eb-ccaa-48ae-8950-2c70e195da03",
-   "metadata": {},
-   "source": [
-    "# Check for Equality between Two MultiIndexSet Instances"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "9a31baab-58d9-4bec-ad3c-9a636b9821b0",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "fb3c2a13-274b-463a-a99e-03b0a0ac3408",
-   "metadata": {},
-   "source": [
-    "Two instances of `MultiIndexSet` may be checked for equality in value via the `==` operator as demonstrated in this guide.\n",
-    "\n",
-    "Two instances of `MultiIndexSet` are equal in value if and only if:\n",
-    "\n",
-    "- The multi-indices elements stored in the `exponents` property of the two instances are equal _and_\n",
-    "- The underlying $l_p$-degrees of the two instances are equal."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "4c6a6b9b-e11f-4f27-96bc-81076bfc37c1",
-   "metadata": {},
-   "source": [
-    "## Equality in value\n",
-    "\n",
-    "Consider the following three-dimensional multi-index sets defined for $l_p$-degree $1.0$:\n",
-    "\n",
-    "$$\n",
-    "A = \\left\\{ (0, 0, 0), (1, 0, 0), (2, 0, 0), (0, 1, 0), (0, 0, 1) \\right\\}\n",
-    "$$\n",
-    "\n",
-    "and create two instances of `MultiIndexSet` with identical exponents and $l_p$-degree."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "b1c866ee-3ef5-4228-b3fc-d4b50993221f",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "mi_1 = mp.MultiIndexSet(\n",
-    "    np.array([[0, 0, 0], [1, 0, 0], [2, 0, 0], [0, 1, 0], [0, 0, 1]]),\n",
-    "    lp_degree=1.0,\n",
-    ")\n",
-    "mi_2 = mp.MultiIndexSet(\n",
-    "    np.array([[0, 0, 0], [1, 0, 0], [2, 0, 0], [0, 1, 0], [0, 0, 1]]),\n",
-    "    lp_degree=1.0,\n",
-    ")"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "1b311986-ea5b-4451-9baa-ca738550c3f0",
-   "metadata": {},
-   "source": [
-    "To check the equality between these two instances, use the `==` operator as follows:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "cc21e7a1-1943-44a3-8c34-1a946507c6e0",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "mi_1 == mi_2"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "12c24ef3-552f-4bc5-afbb-7636a81f5b0a",
-   "metadata": {},
-   "source": [
-    "Note that the two instances are not identical as shown by the following check:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "825c7f8d-bf26-4930-8d58-7f03d151b927",
[…]
