+++ pr-preview/35/_sources/how-to/grid/grid-multiply.ipynb.txt
@@ -1,261 +1,0 @@
-{
- "cells": [
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": "# Multiply `Grid` Instances",
-   "id": "ebf1ed09291f1ec4"
-  },
-  {
-   "cell_type": "code",
-   "id": "initial_id",
-   "metadata": {
-    "collapsed": true
-   },
-   "source": [
-    "import matplotlib.pyplot as plt\n",
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ],
-   "outputs": [],
-   "execution_count": null
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "This guide demonstrates how to multiply instances of `Grid` and the expected outcome of such an operation. The multiplication between two instances of `Grid` may be carried out via the operator `*`.\n",
-    "\n",
-    "The result of a multiplication between two instances of `Grid` is another instance of `Grid` whose underlying multi-index set is the product of the index sets of the operands. This `Grid` instance\n",
-    "can support interpolation polynomials that are the product of polynomials that live on either `Grid` operands.\n",
-    "\n",
-    "To know more about the multi-index set multiplication in Minterpy, see the {doc}`relevant example </how-to/multi-index-set/multi-index-set-multiply>`.\n",
-    "\n",
-    "When multiplying two instances of `Grid` note the following caveats:\n",
-    "\n",
-    "- If a generating function is defined in both instances, then the instances can be multiplied only if the generating functions are compatible, i.e., they are the same function.\n",
-    "  defined in both instances.\n",
-    "- Without a generating function, two instances can be multiplied only if the generating\n",
-    "  points are compatible, i.e., they have the same values up to a common dimension (or columns)."
-   ],
-   "id": "8eee69b4e0b6b159"
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "## Example: Instances with the same dimension\n",
-    "\n",
-    "As a motivating example, consider multiplying two two-dimensional interpolation grids:\n",
-    "\n",
-    "- The first grid has a complete multi-index set with respect to a polynomial degree $2$ and $l_p$-degree $2.0$\n",
-    "- the second grid has a complete multi-index set with respect to a polynomial degree $3$ and $l_p$-degree $1.0$.\n",
-    "\n",
-    "Both interpolation grids have points according to the Leja-ordered Chebyshev-Lobatto points (the default in Minterpy)."
-   ],
-   "id": "c35b5937514ac341"
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": "First, create the instances of `Grid` according to the above specification:",
-   "id": "a957b70b132d8a14"
-  },
-  {
-   "metadata": {},
-   "cell_type": "code",
-   "source": [
-    "grd_1 = mp.Grid.from_degree(2, 2, 2.0)\n",
-    "grd_2 = mp.Grid.from_degree(2, 3, 1.0)"
-   ],
-   "id": "db1a4a8e1f5a1401",
-   "outputs": [],
-   "execution_count": null
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": "The unisolvent nodes that correspond to these are shown below:",
-   "id": "cce000f1c44447d0"
-  },
-  {
-   "metadata": {},
[…]
