+++ pr-preview/35/_sources/how-to/multi-index-set/multi-index-set-downward-closed.ipynb.txt
@@ -1,263 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "5741d863-bc8c-4011-8d63-712c89ef9dd8",
-   "metadata": {},
-   "source": [
-    "# Make a Multi-Index Set Downward-Closed"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "c94d29f1-8250-4231-af69-440ef9df77df",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "79490875-cdde-497a-a5a8-c20588b5ece9",
-   "metadata": {},
-   "source": [
-    "A multi-index set is said to be _downward-closed_ if the set does not contain any lexicographical \"holes\". For a more rigorous definition of downward-closedness, please refer to the {doc}`relevant section </fundamentals/polynomial-bases>` in the docs.\n",
-    "\n",
-    "This guide shows how to check a given `MultiIndexSet` instance for downward-closedness and make the corresponding set of exponents downward-closed."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "d6f277a3-1a78-496f-9154-a092dbded327",
-   "metadata": {},
-   "source": [
-    "## Motivating example\n",
-    "\n",
-    "Consider a non-downward-closed two-dimensional multi-index set with two elements:\n",
-    "\n",
-    "$$\n",
-    "A = \\{ (1, 0), (1, 2) \\}.\n",
-    "$$\n",
-    "\n",
-    "This set is not downward-closed because it contains lexicographical \"holes\" between $(0, 0)$ and the lexicographically largest element $(1,2)$ (e.g., $(0, 0)$ itself, $(0, 1)$)."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "63a9fc78-d3e7-463a-9b11-cedf391ee827",
-   "metadata": {},
-   "source": [
-    "This multi-index set can be created in Minterpy as follows:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "3bff2feb-3a96-4f93-94b7-4ab900c8e4dd",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "mi = mp.MultiIndexSet(np.array([[1, 0], [1, 2]]), lp_degree=1.0)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "42a2bc8a-65db-4177-ad8f-88f912adb019",
-   "metadata": {},
-   "source": [
-    "```{note}\n",
-    "As noted in {doc}`multi-index-set-create`, creating a `MultiIndexSet` instance from the default constructor requires the parameter `lp_degree` (i.e., the $l_p$-degree) to be specified. Here the value of $1.0$ is chosen arbitrarily.\n",
-    "```"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "9a115ec2-1bcc-4dd5-9dc4-c1b5af2c1fdd",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "print(mi)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "f8eacd5f-3ee6-4f29-ba4f-dc325b5a1b42",
-   "metadata": {},
-   "source": [
-    "## Check for downward-closedness\n",
-    "\n",
[…]

[…]
