+++ pr-preview/36/_sources/api/polynomials/canonical.rst.txt
@@ -1,24 +1,0 @@
-====================
-canonical_polynomial
-====================
-
-.. currentmodule:: minterpy.polynomials.canonical_polynomial
-
-.. automodule:: minterpy.polynomials.canonical_polynomial
-
-.. autoclass:: CanonicalPolynomial
-   :members:
-   :show-inheritance:
-   :private-members:
-   :inherited-members:
-   :special-members:
-
-   .. rubric:: Properties
-
-   .. classautosummary:: minterpy.polynomials.canonical_polynomial.CanonicalPolynomial
-      :properties:
-
-   .. rubric:: Methods
-
-   .. classautosummary:: minterpy.polynomials.canonical_polynomial.CanonicalPolynomial
-      :methods:
