+++ pr-preview/36/_sources/contributors/contrib-docs/docs-ipynb.rst.txt
@@ -1,258 +1,0 @@
-#########################
-Writing Jupyter Notebooks
-#########################
-
-The functionalities of Minterpy are demonstrated using documents
-written as Jupyter notebooks, which are executed and integrated into
-the official documentation.
-These built documents can be accessed in the :doc:`/how-to/index`
-and :doc:`/getting-started/index` sections of the Minterpy documentation.
-
-This page outlines the specific guidelines and considerations
-for writing a Jupyter notebook that will be included
-as part of the Minterpy documentation.
-
-Source organization
-===================
-
-Jupyter-notebook-based documents for the Minterpy documentation are currently
-restricted to the following sections:
-
-- :doc:`/getting-started/index`
-- :doc:`/how-to/index`
-
-The source files for these sections are located in the following directories:
-
-- ``/docs/getting-started`` for the :doc:`/getting-started/index`
-- ``/docs/how-to`` for the :doc:`/how-to/index`
-
-Additionally, the How-To Guides are generally organized into a further level
-of hierarchy, based on either:
-
-- **Main Components of Minterpy**, for example, :doc:`/how-to/grid/index`, :doc:`/how-to/multi-index-set/index`.
-- **Numerical Tasks**, for example, :doc:`/how-to/integration/index`.
-
-New tutorials should be put in ``/docs/getting-started`` directory and
-new How-To Guides should be put in a relevant sub-directory of
-``/docs/how-to`` .
-
-Adding a notebook to the source tree
-====================================
-
-If you're adding a new notebook file to the documentation source tree,
-please use one of these locations:
-
-.. rubric:: Getting Started Guides / Tutorials
-
-Place new Tutorials in the ``/docs/getting-started`` directory.
-
-.. rubric:: How-To Guides
-
-Place new How-To Guides in the appropriate subdirectory within ``/docs/how-to``,
-depending on the topic.
-
----
-
-By adhering to this structure, the Jupyter-notebook-based documents will be
-correctly organized and accessible within the Minterpy documentation.
-
-Adding a notebook to documentation
-==================================
-
-If you add a new Jupyter notebook to the Minterpy documentation,
-ensure that the relevant ``index.rst`` files are updated accordingly.
-
-.. rubric:: Updating the index.rst File
-
-In the appropriate ``index.rst`` file, you must add a new entry to
-the ``toctree`` directive to include the new notebook.
-This ensures the new document is accessible in the built documentation.
-
-.. code-block:: rest
-
-   .. toctree::
-      :maxdepth: 1
-      :hidden:
-
-      ...
-
-.. rubric:: Getting Started Guides / Tutorials
-
-For new tutorials, update the ``/docs/getting-started/index.rst`` file
-with the new notebook entry under the ``toctree`` directive.
-
-.. rubric:: How-To Guides:
-
-If the new How-To Guide belongs to an existing sub-directory,
-update the corresponding ``/docs/how-to/<relevant-directory>/index.rst`` file.
-
-If it does not belong to an existing sub-directory, follow these steps:
-
-- Create a new directory for the guide with its own ``index.rst`` file.
-- Put the new notebook file in the new directory.
-- Add a ``toctree`` directive inside the new ``index.rst`` to list the new file.
-- Update the parent ``/docs/how-to/index.rst`` to include the new directory
-  and its index file.
-
-----
-
-By following these steps, you ensure that the new Jupyter notebooks are properly linked and accessible within the Minterpy documentation structure.
-
-Setup
-=====
-
-We recommend creating and editing Jupyter notebooks using JupyterLab,
-which is included as part of the extras requirement
-for the documentation development environment.
-
-If you follow the installation steps in
[…]
