+++ latest/_sources/getting-started/md-polynomial-interpolation.ipynb.txt
@@ -595,7 +595,7 @@
     "Q(\\boldsymbol{x}) = \\sum_{\\boldsymbol{\\alpha} \\in A_{m, n, p}} c_{\\mathrm{lag}, \\boldsymbol{\\alpha}} \\Psi_{\\mathrm{lag}, \\boldsymbol{\\alpha}}(\\boldsymbol{x}),\n",
     "$$\n",
     "\n",
-    "where $A_{m, n, p}$ is the complete multi-index set and $\\Psi_{\\mathrm{lag}, \\boldsymbol{\\alpha}}$'s are the monomials of the Lagrange basis that satisfy:\n",
+    "where $A_{m, n, p}$ is the complete multi-index set and $\\Psi_{\\mathrm{lag}, \\boldsymbol{\\alpha}}$'s are the Lagrange basis polynomials, each of which satisfies:\n",
     "\n",
     "$$\n",
     "L_{\\boldsymbol{\\alpha}}(p_{\\boldsymbol{\\beta}}) = \\delta_{\\boldsymbol{\\alpha}, \\boldsymbol{\\beta}}, p_{\\boldsymbol{\\beta}} \\in P_{A_{m, n, p}},\n",
@@ -641,7 +641,7 @@
     "Q(\\boldsymbol{x}) = \\sum_{\\boldsymbol{\\alpha} \\in A_{m, n, p}} c_{\\mathrm{nwt}, \\boldsymbol{\\alpha}} \\, \\Psi_{\\mathrm{nwt}, \\boldsymbol{\\alpha}}(\\boldsymbol{x}),\n",
     "$$\n",
     "\n",
-    "where $c_{\\mathrm{nwt}, \\boldsymbol{\\alpha}}$'s are the coefficients of the polynomial in the multidimensional Newton basis and $\\Psi_{\\mathrm{nwt}, \\boldsymbol{\\alpha}}$'s are the monomials of the Newton basis. The monomial associated with multi-index element $\\boldsymbol{\\alpha}$ is defined as follows:\n",
+    "where $c_{\\mathrm{nwt}, \\boldsymbol{\\alpha}}$'s are the coefficients of the polynomial in the multidimensional Newton basis and $\\Psi_{\\mathrm{nwt}, \\boldsymbol{\\alpha}}$'s are the Newton basis polynomials. The basis polynomial associated with multi-index element $\\boldsymbol{\\alpha}$ is defined as follows:\n",
     "\n",
     "$$\n",
     "\\Psi_{\\mathrm{nwt}, \\boldsymbol{\\alpha}}(\\boldsymbol{x}) = \\prod_{j = 1}^m \\prod_{i = 0}^{\\alpha_j - 1} (x_j - p_{i, j}),\n",
