+++ pr-preview/37/_sources/api/polynomials/chebyshev.rst.txt
@@ -1,0 +1,24 @@
+====================
+chebyshev_polynomial
+====================
+
+.. currentmodule:: minterpy.polynomials.chebyshev_polynomial
+
+.. automodule:: minterpy.polynomials.chebyshev_polynomial
+
+.. autoclass:: ChebyshevPolynomial
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.polynomials.chebyshev_polynomial.ChebyshevPolynomial
+      :properties:
+
+   .. rubric:: Methods
+
+   .. classautosummary:: minterpy.polynomials.chebyshev_polynomial.ChebyshevPolynomial
+      :methods:
