+++ pr-preview/37/_sources/api/core/grid.rst.txt
@@ -1,0 +1,24 @@
+====
+grid
+====
+
+.. currentmodule::minterpy.core.grid
+
+.. automodule:: minterpy.core.grid
+
+.. autoclass:: Grid
+    :members:
+    :private-members:
+    :inherited-members:
+    :special-members:
+    :exclude-members: __init__
+
+    .. rubric:: Properties
+
+    .. classautosummary:: minterpy.core.grid.Grid
+       :properties:
+
+    .. rubric:: Methods
+
+    .. classautosummary:: minterpy.core.grid.Grid
+       :methods:
