+++ pr-preview/37/_sources/api/transformation-operators/operator-barycentric/operator-barycentric.rst.txt
@@ -1,24 +1,0 @@
-===================
-BarycentricOperator
-===================
-
-.. automodule:: minterpy.schemes.barycentric.operators
-
-.. currentmodule:: minterpy.schemes.barycentric.operators
-
-.. autoclass:: BarycentricOperator
-   :members:
-   :show-inheritance:
-   :private-members:
-   :inherited-members:
-   :special-members:
-
-   .. rubric:: Properties
-
-   .. classautosummary:: minterpy.schemes.barycentric.operators.BarycentricOperator
-      :properties:
-
-   .. rubric:: Methods
-
-   .. classautosummary:: minterpy.schemes.barycentric.operators.BarycentricOperator
-      :methods:
