+++ pr-preview/37/_sources/how-to/integration/index.rst.txt
@@ -1,13 +1,0 @@
-###########
-Integration
-###########
-
-This page lists all the available How-To Guides related to
-polynomial integration.
-
-.. toctree::
-   :maxdepth: 1
-
-   Integrate a One-Dimensional Polynomial <poly-interpolant-integration-1d>
-   Integrate an M-Dimensional Polynomial <poly-interpolant-integration-md>
-
