+++ pr-preview/37/_sources/getting-started/arithmetic-operations-with-polynomials.ipynb.txt
@@ -1,1163 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "60f73f47",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "# Basic Arithmetic Operations with Polynomials"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "9250bb6bcb0da956",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "Once you have constructed a polynomial in Minterpy, you can perform various operations on it, depending on the basis in which the polynomial is represented.\n",
-    "For instance, as demonstrated in the previous tutorials, you can evaluate a Minterpy polynomial (in the Newton basis) at a set of query points.\n",
-    "\n",
-    "This tutorial will show you how to extend beyond evaluation by performing arithmetic operations with Minterpy polynomials.\n",
-    "Specifically, you'll learn that Minterpy polynomials can be:\n",
-    "\n",
-    "- Added or subtracted from other polynomials\n",
-    "- Multiplied by other polynomials or real scalar numbers\n",
-    "- Raised to a power (as long as it's a non-negative integer)\n",
-    "\n",
-    "These operations result in another Minterpy polynomial in the same basis.\n",
-    "In other words, Minterpy polynomials are **closed** under the following arithmetic operations:\n",
-    "\n",
-    "- Multiplying a polynomial by another polynomial\n",
-    "- Multiplying a polynomial by a scalar (a real number)\n",
-    "- Adding or subtracting one polynomial to/from another\n",
-    "- Adding or subtracting a scalar (a real number) to/from a polynomial\n",
-    "- Raising a polynomial to a non-negative integer power\n",
-    "\n",
-    "To keep things simple, this tutorial uses one-dimensional polynomials as examples. However, the principles and behaviors we'll cover apply to Minterpy polynomials of any dimensions.\n",
-    "\n",
-    "Before you start, make sure to import the necessary packages to follow along with this guide."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "aa387b47-58fc-453c-b738-662bc24ff0fc",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np\n",
-    "import matplotlib.pyplot as plt"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "52ea8534-a3d0-4b98-807b-25777b4594f0",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "## Motivating functions\n",
-    "\n",
-    "Consider the following two one-dimensional functions.\n",
-    "\n",
[…]
