+++ pr-preview/37/_sources/contributors/contrib-docs/docs-design.rst.txt
@@ -1,186 +1,0 @@
-=====================
-Documentation: Design
-=====================
-
-We believe that documentation is essential to the on-going development of
-any software.
-Good documentation helps users and contributors--including ourselves--navigate
-the steep learning curve of using and contribution to Minterpy.
-That's why we invest significant time in developing
-and improving the documentation.
-
-On this page, you'll learn more about the underlying design of the Minterpy
-documentation.
-
-Design principles
-=================
-
-There isn't just *one* documentation but *four*,
-each serving a distinct purpose.
-We follow this documentation design principles from `Diátaxis`_
-to structure the Minterpy documentation.
-The four different *categories* of documentation are:
-
-- :ref:`contributors/contrib-docs/docs-design:Tutorials`
-- :ref:`contributors/contrib-docs/docs-design:How-to guides`
-- :ref:`contributors/contrib-docs/docs-design:Explanation (fundamentals, theory)`
-- :ref:`(API) reference <contributors/contrib-docs/docs-design:API reference>`
-
-The distinction between these categories is summarized in the table below.
-
-===============  ===============  =============== ===================== ===========
-Category         Orientation      Main content    Most useful when...   Think of...
-===============  ===============  =============== ===================== ===========
-Tutorials        Learning         Practical steps learning the code     Lessons
-How-to           Problem-solving  Practical steps working with the code Recipes
-Explanation      Understanding    Exposition      learning the code     Textbooks
-(API) reference  Information      Exposition      working with the code Dictionary
-===============  ===============  =============== ===================== ===========
-
-Or if you're more visually-oriented...
-
-.. figure:: /assets/images/contributors/documentation-system.png
-  :align: center
-
-  Four distinct categories of documentation (adapted from the `Diátaxis`_).
-
-.. seealso::
-
-   We encourage you to read through the design as outlined in
-   the `Diátaxis`_ and watch a talk by the author Daniele Procida
-   in `Pycon 2017`_.
-
-The :doc:`main sections <docs-source-structure>`
-of the Minterpy documentation directly reflect all these four categories.
-Additionally, we have a special section for contributors called
-the :doc:`/contributors/index`, which primarily contains how-to guides tailored
-for contributors rather than users.
-Each addition to the documentation should be clearly assigned to a single
-category.
-
-Tutorials
-=========
-
-*Tutorials* are *learning-oriented* documents; think of them as *lessons*
-for Minterpy users.
-In a tutorial, the tasks should be meaningful and the outcomes should be immediate,
-while avoiding unecessary explanations, jargon, and technical details--especially
-in the beginning.
-They should have a minimal, but non-trivial, context.
-Writing tutorials is challenging; you need to select a problem that is
-relatively easy yet non-trivial
-and carefully introduce important and more advanced topics as the tutorial
-progresses.
-
-Well-curated tutorials are a crucial, if not the most important,
-element of the documentation for onboarding both users and contributors.
-The offer a glimpse of what Minterpy can do and how to use it as intended.
-
-We organize all tutorials inside the :doc:`/getting-started/index` and list
-them in order of difficulty. This section includes a quickstart guide
-and a series of in-depth tutorials.
-
-How-to guides
-=============
-
-*How-to Guides* are *problem-solving-oriented* documents;
-think of them as *recipes* for performing specific tasks with Minterpy.
-Each guide provides step-by-step instructions for solving a clearly defined problem.
-These guides are largely context-free, assuming that users who have some
-proficiency with Minterpy are seeking solutions to common, specific issues.
-
-Writing how-to guides is somewhat easier than writing tutorials, as the problems
[…]

[…]
