+++ pr-preview/38/_sources/getting-started/polynomial-regression.ipynb.txt
@@ -1,0 +1,934 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "id": "18832584-2604-4143-b435-d224d8fcc778",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "# Polynomial Regression with Minterpy"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "60effc2d-4ece-438e-ad0f-3d683703d964",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
+    "The polynomials you've worked with in these tutorials are designed to approximate functions of interest by using values at specific, strategically chosen points that form unisolvent nodes.\n",
+    "This implies that you have the flexibility to evaluate the function at these selected points within its domain.\n",
+    "\n",
+    "However, situations may arise where function values are available only at arbitrary, scattered points across the domain.\n",
+    "In such cases, you cannot select the evaluation points freely.\n",
+    "Despite this limitation, it is still possible to derive a polynomial that best fits the scattered data using the least squares method.\n",
+    "\n",
+    "In this in-depth tutorial, you'll learn how to perform polynomial regression in Minterpy to construct a polynomial that fits scattered data. This process involves finding the polynomial that minimizes the error between the observed values and the values predicted by the polynomial."
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "id": "9d5175a7-c1c6-4095-97b2-30aa2fc5f8b0",
+   "metadata": {
+    "editable": true,
+    "slideshow": {
+     "slide_type": ""
+    },
+    "tags": []
+   },
+   "source": [
[…]
