+++ pr-preview/38/_sources/contributors/about-us.rst.txt
@@ -1,0 +1,156 @@
+========
+About Us
+========
+
+Minterpy is an optimized Python implementation of
+the multidimensional polynomial interpolation prototype algorithm called MIP,
+which was originally developed in MATLAB
+by Hecht et al.\ :footcite:`Hecht2020`.
+
+It is developed and maintained by a small team at the Center
+for Advanced Systems Understanding (`CASUS`_), an institute of the
+Helmholtz-Zentrum Dresden-Rossendorf (`HZDR`_).
+
+.. raw:: html
+
+   <table>
+      <tr>
+         <td align="center"><a href="https://www.casus.science/"><img src="../_static/CASUS.png" width="250px;" alt="CASUS"/></td>
+         <td align="center"><a href="https://www.hzdr.de/"><img src="../_static/HZDR.png" width="250px;" alt="HZDR"/></td>
+      </tr>
+   </table>
+
+Maintainers and contributors
+============================
+
+.. rubric:: Project maintainers
+
+Project maintainers are the current main developers responsible for Minterpy
+and managing its development process.
+
+.. raw:: html
+
+   <!-- ALL-CONTRIBUTORS-LIST:START - Do not remove this section -->
+   <table>
+      <tr>
+         <td align="center"><a href="https://www.casus.science/team-members/uwe-hernandez-acosta/"><img src="https://gitlab.hzdr.de/uploads/-/system/user/avatar/88/avatar.png?width=90" width="100px;" alt=""/><br /><sub><b>Uwe Hernandez Acosta</b></sub></a><br />
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/merge_requests?scope=all&state=all&assignee_username=hernan68" title="Maintenance">🚧</a>
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/merge_requests?scope=all&state=all&author_username=hernan68" title="Code">💻</a>
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/issues?scope=all&state=all&author_username=hernan68" title="Bug reports">🐛</a>
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/merge_requests?scope=all&state=all&reviewer_username=hernan68" title="Reviewed Merge Requests">👀</a><br />
+            <a href="#ideas-hernan68" title="Ideas, Planning, & Feedback">🤔</a>
+         </td>
+         <td align="center"><a href="https://www.casus.science/team-members/dr-damar-wicaksono/"><img src="https://de.gravatar.com/userimage/108859669/c1851123ff56c230cf2a67834ee1337b.jpeg" width="100px;" alt=""/><br /><sub><b>Damar Wicaksono</b></sub></a><br />
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/merge_requests?scope=all&state=all&assignee_username=wicaks72" title="Maintenance">🚧</a>
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/merge_requests?scope=all&state=all&author_username=wicaks72" title="Code">💻</a>
+            <a href="https://gitlab.hzdr.de/interpol/minterpy/-/issues?scope=all&state=all&author_username=wicaks72" title="Bug reports">🐛</a><br />
+            <a href="https://gitlab.hzdr.de/dashboard/merge_requests?scope=all&state=all&reviewer_username=wicaks72" title="Reviewed Merge Requests">👀</a>
[…]
