+++ pr-preview/38/_sources/how-to/multi-index-set/multi-index-set-empty.ipynb.txt
@@ -1,514 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "d6a11f9a-d895-4cc7-befa-bac340b8915d",
-   "metadata": {},
-   "source": [
-    "# Create an Empty Multi-Index Set"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 1,
-   "id": "b86f737a-bdc7-432e-ab3f-939491391220",
-   "metadata": {
-    "ExecuteTime": {
-     "end_time": "2024-02-20T14:32:48.002135200Z",
-     "start_time": "2024-02-20T14:32:47.034693500Z"
-    }
-   },
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "b81abe18-3ed4-4925-a09e-bdd95e4b0880",
-   "metadata": {},
-   "source": [
-    "An empty `MultiIndexSet` contains an empty array as its exponents. Such a set may arise from some set operations involving `MultiIndexSet`.\n",
-    "\n",
-    "This guide demonstrates how to create such an array and the corresponding conventions."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "c05807a1-0c6e-4d6c-ae54-2e4ee89a28e4",
-   "metadata": {},
-   "source": [
-    "## Construction\n",
-    "\n",
-    "An empty `MultiIndexSet` instance may be constructed by passing an empty two-dimensional array as the exponents to the default constructor:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 2,
-   "id": "5a556b62-e39a-4b85-a9f2-2aa78534bbc6",
-   "metadata": {
-    "ExecuteTime": {
-     "end_time": "2024-02-20T14:32:48.021146600Z",
-     "start_time": "2024-02-20T14:32:48.003145200Z"
-    }
-   },
-   "outputs": [
-    {
-     "name": "stdout",
-     "output_type": "stream",
-     "text": [
-      "MultiIndexSet(m=0, n=None, p=1.0)\n",
-      "[]\n"
-     ]
-    }
-   ],
-   "source": [
-    "exponent = np.empty((0, 0), dtype=int)\n",
-    "mi = mp.MultiIndexSet(exponent, lp_degree=1.0)\n",
-    "print(mi)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "40802d6d-13e8-4334-a148-75299ca99c00",
-   "metadata": {},
-   "source": [
-    "Note that just like any other `MultiIndexSet` instances, the $l_p$-degree must be specified during the construction."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "0380be35-2e94-4079-976e-bbc99af6bb12",
-   "metadata": {},
-   "source": [
-    "The cardinality of an empty set is 0:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 3,
-   "id": "fc248000-6cd2-4187-b8b4-885f2ef12a0d",
-   "metadata": {
-    "ExecuteTime": {
-     "end_time": "2024-02-20T14:32:48.052134700Z",
-     "start_time": "2024-02-20T14:32:48.021146600Z"
-    }
-   },
-   "outputs": [
-    {
-     "data": {
-      "text/plain": "0"
-     },
-     "execution_count": 3,
-     "metadata": {},
-     "output_type": "execute_result"
-    }
-   ],
-   "source": [
-    "len(mi)"
-   ]
-  },
-  {
[…]
