+++ pr-preview/38/_sources/api/transformations/transformation-identity.rst.txt
@@ -1,19 +1,0 @@
-========
-identity
-========
-
-.. currentmodule:: minterpy.transformations.identity
-
-.. automodule:: minterpy.transformations.identity
-
-.. autoclass:: Identity
-   :members:
-   :show-inheritance:
-   :private-members:
-   :inherited-members:
-   :special-members:
-
-   .. rubric:: Properties
-
-   .. classautosummary:: minterpy.transformations.identity.Identity
-      :properties:
