+++ pr-preview/38/_sources/api/core/ABC/abc-operator.rst.txt
@@ -1,18 +1,0 @@
-=================
-operator_abstract
-=================
-
-.. currentmodule:: minterpy.core.ABC.operator_abstract
-
-.. automodule:: minterpy.core.ABC.operator_abstract
-
-.. autoclass:: OperatorABC
-   :members:
-   :private-members:
-   :inherited-members:
-   :special-members: __matmul__
-
-   .. rubric:: Properties
-
-   .. classautosummary:: minterpy.core.ABC.operator_abstract.OperatorABC
-      :properties:
