+++ joss/paper.bib
@@ -325,8 +325,8 @@
   author = {Wicaksono, Damar Canggih and Hernandez Acosta, Uwe and Thekke Veettil, Sachin Krishnan and Kissinger, Jannik and Hecht, Michael},
   title  = {{Data to "Minterpy: Multivariate polynomial interpolation in Python"}},
   year   = {2025},
-  doi    = {10.14278/rodare.3379},
-  url    = {https://doi.org/10.14278/rodare.3379},
+  doi    = {10.14278/rodare.3717},
+  url    = {https://doi.org/10.14278/rodare.3717},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
