+++ .github/workflows/draft-pdf.yml
@@ -1,5 +1,8 @@
 name: Draft PDF
-on: [push]
+on:
+  push:
+    branches:
+      - dev-joss
 
 jobs:
   paper:
