### 2025-04-24 #7702 @damar-wicaksono
@sixpearls,

Thank you for the additional comments:

> Fix the math rendering error on line 101

This has been fixed (it's now line 99).

> Provide some mention of the tradeoff of run-time for the increased accuracy and convenience for certain mathematical operations. The global polynomial requires many more floating point multiplications and additions compared to piecewise polynomials (e.g., the compact support of B-splines).

This is indeed an important point; we've added some comments on this to the draft manuscript. You can jump to line 149-158 to check it out. I'll ask the editorial bot to regenerate the draft manuscript below.

Thank you very much again for your feedbacks! Please let me know if you have any further comments.
