+++ pr-preview/39/_sources/api/internal/utils/arrays.rst.txt
@@ -1,6 +1,0 @@
-======
-arrays
-======
-
-.. automodule:: minterpy.utils.arrays
-   :members:
