### 2025-04-25 c9352e0
Revert " Fix stale information on the contributors guide"
