+++ joss/paper.bib
@@ -1,332 +1,0 @@
-@Article{Dornheim2023,
-  author    = {Dornheim, Tobias and Wicaksono, Damar and Suarez-Cardona, Juan E. and Tolias, Panagiotis and Böhme, Maximilian P. and Moldabekov, Zhandos A. and Hecht, Michael and Vorberger, Jan},
-  journal   = {Physical Review B},
-  title     = {Extraction of the frequency moments of spectral densities from imaginary-time correlation function data},
-  year      = {2023},
-  number    = {15},
-  pages     = {155148},
-  volume    = {107},
-  doi       = {10.1103/physrevb.107.155148},
-  publisher = {American Physical Society (APS)},
-}
-
-@InProceedings{Schreiber2023,
-  author    = {Schreiber, Janina and Wicaksono, Damar and Hecht, Michael},
-  booktitle = {Proceedings of the Companion Conference on Genetic and Evolutionary Computation},
-  title     = {Minimizing black boxes due to polynomial-model-based optimization},
-  year      = {2023},
-  pages     = {759--762},
-  series    = {GECCO ’23 Companion},
-  volume    = {2},
-  doi       = {10.1145/3583133.3590743},
-}
-
-@Article{Veettil2023,
-  author  = {Veettil, Sachin Krishnan Thekke and Zavalani, Gentian and Acosta, Uwe Hernandez and Sbalzarini, Ivo F. and Hecht, Michael},
-  journal = {SIAM Journal on Scientific Computing},
-  title   = {Global polynomial level sets for numerical differential geometry of smooth closed surfaces},
-  year    = {2023},
-  number  = {4},
-  pages   = {A1995--A2018},
-  volume  = {45},
-  doi     = {10.1137/22m1536510},
-}
-
-@Article{Meijering2002,
-  author  = {Meijering, E.},
-  journal = {Proceedings of the IEEE},
-  title   = {A chronology of interpolation: from ancient astronomy to modern signal and image processing},
-  year    = {2002},
-  number  = {3},
-  pages   = {319--342},
-  volume  = {90},
-  doi     = {10.1109/5.993400},
-}
-
-@Article{Cools2002,
-  author  = {Cools, Ronald},
-  journal = {Journal of Computational and Applied Mathematics},
-  title   = {Advances in multidimensional integration},
-  year    = {2002},
-  number  = {1},
-  pages   = {1--12},
-  volume  = {149},
-  doi     = {10.1016/s0377-0427(02)00517-4},
-}
-
-@Article{Xiu2009,
-  author  = {Xiu, Dongbin},
-  journal = {Communications in Computational Physics},
-  title   = {Fast numerical methods for stochastic computations: a review},
-  year    = {2009},
-  number  = {2--4},
-  pages   = {242--272},
-  volume  = {5},
-}
-
-@Book{Trefethen2019,
-  author    = {Trefethen, L. N.},
-  publisher = {Society for Industrial and Applied Mathematics (SIAM)},
-  title     = {Approximation theory and approximation practice, extended edition},
-  year      = {2019},
-  address   = {Philadelphia, PA},
-}
-
-@Article{Branges1959,
-  author  = {{de Branges}, Louis},
-  journal = {Proceedings of the American Mathematical Society},
-  title   = {The {Stone}-{Weierstrass} theorem},
-  year    = {1959},
-  number  = {5},
-  pages   = {822--824},
-  volume  = {10},
-  doi     = {10.1090/s0002-9939-1959-0113131-7},
-}
-
-@Article{Trefethen2011,
-  author  = {Trefethen, L. N.},
[…]
