+++ CONTRIBUTING.md
@@ -1,38 +1,36 @@
 # Contribution Guide
 
-Thanks a lot for your interest and taking the time to contribute
-to the Minterpy project!
+Thanks a lot for your interest and taking the time to contribute to the `minterpy` project!
 
-This document provides short guidelines for contributing
-to the Minterpy project.
-For a more comprehensive guide, please refer to the online documentation
-([stable](https://minterpy-project.github.io/minterpy/stable/contributors/index.html)
-or [latest](https://minterpy-project.github.io/minterpy/latest/contributors/index.html))
+This document provides guidelines for contributing to the `minterpy` project.
 
 ## Installation
 
 This installation guide is focused on development.
-For installing Minterpy in production environment, check out [README.md](./README.md).
+For installing `minterpy` in production runs check out the [README.md](./README.md).
 
-### Obtaining the source
-
-To obtain the latest source,
-clone the Minterpy repository from [GitHub](https://github.com/minterpy-project/minterpy):
+In order to get the source of latest release,
+clone the `minterpy` repository from the [HZDR GitLab]:
 
 ```bash
-git clone https://github.com/minterpy-project/minterpy
+git clone https://gitlab.hzdr.de/interpol/minterpy.git
 ```
 
-By default, the cloned branch is the `dev` branch (i.e., the latest development
-version).
+By default, the cloned branch is the `main` branch.
 
-We recommend always pulling the latest commit:
+To get the latest development version, checkout to the `dev` branch:
+
+```bash
+git checkout dev
+```
+
+We recommend to always pull the latest commit:
 
 ```bash
 git pull origin dev
 ```
 
-You are not allowed to directly push to `dev` or `main` branch.
+You are not allowed to directly push to `dev` or `master` branch.
 Please follow the instructions under [Branching workflow](#branching-workflow).
 
 ### Virtual environments
@@ -41,7 +39,7 @@ Following a best practice in Python development,
 we strongly encourage you to create and use virtual environments for development and production runs.
 A virtual environment encapsulates the package and all dependencies without messing up your other Python installations.
 
-The following instructions should be executed from the Minterpy source directory.
+The following instructions should be executed from the `minterpy` source directory.
 
 #### Using [venv](https://docs.python.org/3/tutorial/venv.html) from the python standard library:
 
@@ -51,7 +49,7 @@ The following instructions should be executed from the Minterpy source directory
     python -m venv <your_venv_name>
     ```
 
-   Replace `<your_venv_name>` with an environment name of your choice.
+   Replace `<you_venv_name>` with an environment name of your choice.
 
 2. Activate the environment you just created:
 
@@ -59,9 +57,11 @@ The following instructions should be executed from the Minterpy source directory
     source <your_venv_name>/bin/activate
     ```
 
-    Replace <your_venv_name> with your desired environment name.
+    as before replace `<you_venv_name>` with the environment name.
+
 
 3. To deactivate the virtual environment, type:
+
 
     ```bash
     deactivate
@@ -75,17 +75,18 @@ The following instructions should be executed from the Minterpy source directory
     virtualenv <your_venv_name>
     ```
 
-   Replace `<your_venv_name>` with an environment name of your choice.
+   Replace `<you_venv_name>` with an environment name of your choice.
 
-2. Activate the environment you just created:
+2. Activate the environment:
 
     ```bash
     source <your_venv_name>/bin/activate
     ```
 
-    Replace <your_venv_name> with your desired environment name.
+    As before replace `<you_venv_name>` with the environment name.
 
 3. To deactivate the virtual environment, type:
+
 
     ```bash
     deactivate
@@ -99,18 +100,17 @@ The following instructions should be executed from the Minterpy source directory
[…]
