+++ pr-preview/40/_sources/api/transformation-operators/operator-barycentric/operator-barycentric-piecewise.rst.txt
@@ -1,0 +1,22 @@
+============================
+BarycentricPiecewiseOperator
+============================
+
+.. currentmodule:: minterpy.schemes.barycentric.operators
+
+.. autoclass:: BarycentricPiecewiseOperator
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.schemes.barycentric.operators.BarycentricPiecewiseOperator
+      :properties:
+
+   .. rubric:: Methods
+
+   .. classautosummary:: minterpy.schemes.barycentric.operators.BarycentricPiecewiseOperator
+      :methods:
