+++ pr-preview/40/_sources/how-to/index.rst.txt
@@ -1,0 +1,28 @@
+#############
+How-to Guides
+#############
+
+This page contains a series of guides on how to carry out routine numerical
+tasks with polynomials in Minterpy as well as demonstrating important features
+of the main components of Minterpy.
+
+.. toctree::
+   :maxdepth: 2
+
+   Multi-index Set <multi-index-set/index>
+   Grid <grid/index>
+   Integration <integration/index>
+
+..
+    .. todo::
+
+       This chapter should provide a set of practical examples on how to achieve `common tasks`
+       with ``minterpy``. We can assume that the users have some basic knowledge,
+       they know what they want to achieve, they just don't know how to do it in ``minterpy``.
+       Think of `recipes` when creating an entry in the How-to Guides chapter
+
+       For examples:
+
+       - Surface interpolation
+       - Transformation between bases
+       - Differentiation
