+++ pr-preview/41/_sources/api/internal/jit_compiled/canonical.rst.txt
@@ -1,0 +1,6 @@
+=========
+canonical
+=========
+
+.. automodule:: minterpy.jit_compiled.canonical
+   :members:
