+++ pr-preview/41/_sources/fundamentals/transformation.rst.txt
@@ -1,0 +1,124 @@
+============================
+Transformation Between Bases
+============================
+
+If an :math:`m`-dimensional polynomial :math:`Q \in \Pi_A`,
+where :math:`A \subseteq \mathbb{N}^m`, is given in one of the
+:doc:`supported bases </fundamentals/polynomial-bases>`, you might want
+to transform it into another representation to leverage specific properties.
+For instance, canonical polynomials are familiar while Newton polynomials
+are more stable and accurate to evaluate.
+
+On this page, we outline how the underlying theorem how such a transformation
+is carried out.
+
+Theorem: Basis transformations
+==============================
+
+Let :math:`A= A_{m,n,p}\,, m,n \in \mathbb{N}\,, p > 0`
+be a :ref:`complete multi-index set
+<fundamentals/polynomial-bases:Complete multi-index set>`,
+:math:`P_A \subseteq \Omega` be :ref:`unisolvent nodes
+<fundamentals/interpolation-at-unisolvent-nodes:Unisolvent nodes>`,
+and :math:`Q \in \Pi_A` be a polynomial.
+Then:
+
+- Lower triangular matrices
+  :math:`\mathrm{NL}_A, \mathrm{LN}_A \in \mathbb{R}^{|A|\times |A|}`
+  can be computed in :math:`\mathcal{O}(|A|^3)` operations, such that
+
+  .. math::
+     \mathrm{LN}_A \cdot\mathrm{NL}_A = \mathrm{I} \,, \quad
+     \mathrm{NL}_A \cdot \boldsymbol{c}_{\mathrm{nwt}} =
+     \boldsymbol{c}_{\mathrm{lag}}\,, \quad
+     \mathrm{LN}_A\cdot \boldsymbol{c}_{\mathrm{lag}} =
+     \boldsymbol{c}_{\mathrm{nwt}} \,,
+
+  where :math:`\boldsymbol{c}_{\mathrm{lag}} =
+  (c_{\mathrm{lag}, \boldsymbol{\alpha}})_{\alpha \in A} \in \mathbb{R}^{\lvert A \rvert}`
+  and :math:`\boldsymbol{c}_{\mathrm{nwt}} =
+  (c_{\mathrm{lag}, \boldsymbol{\alpha}})_{\alpha \in A} \in \mathbb{R}^{\lvert A \rvert}`
+  are the Lagrange coefficients and the Newton coefficients of :math:`Q`,
+  respectively.
+
+- Upper triangular matrices
+  :math:`\mathrm{CL}_A, \mathrm{CN}_A \in \mathbb{R}^{\lvert A \rvert \times \lvert A \rvert}`
+  can be computed in :math:`\mathcal{O}(\lvert A \rvert^3)` operations,
+  such that
+
+  .. math::
+     \mathrm{CL}_A \cdot \boldsymbol{c}_{\mathrm{can}} =
+     \boldsymbol{c}_{\mathrm{lag}}\,, \quad \mathrm{CN}_A \cdot
+     \boldsymbol{c}_{\mathrm{can}} = \boldsymbol{c}_{\mathrm{nwt}}\,,
+
+  where :math:`\boldsymbol{c}_{\mathrm{can}} =
+  (c_{\mathrm{can}, \boldsymbol{\alpha}})_{\alpha \in A} \in \mathbb{R}^{\lvert A \rvert}`
+  denotes the canonical coefficients of :math:`Q`.
+
+Due to their triangular structure, the inverse matrices
+:math:`\mathrm{NC}_A =\mathrm{CN}_A^{-1}`,
+:math:`\mathrm{LC}_A =\mathrm{CL}_A^{-1}` can be computed efficiently
[…]
