+++ pr-preview/41/_sources/getting-started/function-approximations.ipynb.txt
@@ -1,1033 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "60f73f47",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "# Quick Start Guide to Function Approximations with Minterpy"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "4b710d24-a1e1-49e0-8816-b62886fa8556",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "Welcome to the Quickstart Guide to Minterpy!\n",
-    "\n",
-    "If you're reading this, chances are you want to create an interpolant that approximates a function, possibly in multiple dimensions. This guide provides a quick walkthrough on how to achieve this using Minterpy."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "7df7d579-94ce-47aa-b682-262df7f192bf",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "## How to import Minterpy\n",
-    "\n",
-    "After installing {ref}`installing Minterpy <getting-started/index:Installation>`, you can import it as follows:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "4c2b0aa0-e94e-45d5-8a28-96309e80b113",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "outputs": [],
-   "source": [
-    "import minterpy as mp"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "6d099ed5-5ee9-4eb9-94ab-5efb1b2e3770",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "The shorthand `mp` is the convention we, the developer team, like to use.\n",
-    "You're free to choose any shorthand you prefer, but we find `mp` to be both short and recognizable (similar to `np` for NumPy and `pd` for pandas).\n",
-    "\n",
-    "While you're at it, you will also need to import Numpy and Matplotlib to follow along with this guide."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "6027a0cb-14e0-4f26-992e-57e54cd5d4ac",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "outputs": [],
-   "source": [
-    "import numpy as np\n",
-    "import matplotlib.pyplot as plt"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "03814b27-6007-405a-96bf-140a0f1ce2c5",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "## One-dimensional function\n",
-    "\n",
-    "Let's start with interpolating the following one-dimensional function:\n",
-    "\n",
-    "$$\n",
-    "f(x) = x \\, \\sin{(12 x)}, x \\in [-1, 1].\n",
-    "$$"
-   ]
-  },
-  {
-   "cell_type": "markdown",
[…]
