+++ pr-preview/41/_sources/contributors/development-environment.rst.txt
@@ -1,430 +1,0 @@
-====================================
-Setting Up a Development Environment
-====================================
-
-Before contributing to Minterpy, we recommend setting up a development
-environment on your systems so you can start developing from the correct branch
-and test your changes or build the documentation.
-
-Forking and cloning the repository
-==================================
-
-Minterpy is hosted on `GitHub`_; to contribute you will need to sign up
-for a free GitHub account.
-
-We use `Git`_ for version control, which opens up a wide range of opportunities
-for people to contribute to Minterpy.
-
-Getting started with Git
-------------------------
-
-If you're new to Git, you can refer to the `Git documentation`_ for some
-learning materials.
-
-Additionally, please refer to the `GitHub documentation`_ for instructions
-on installing Git, setting up your SSH key, and configuring Git.
-Completing these steps is essential for working seamlessly between your local
-repository and GitHub.
-
-Forking the Minterpy repository
--------------------------------
-
-The Minterpy project follows a forking workflow.
-In this model, contributors fork the repository, make changes, and then create
-a merge (pull) request.
-This means that, first and foremost, you need your own fork of the Minterpy
-repository to work on the code.
-
-Go to the Minterpy `repository page`_ and click the **Fork** button.
-Make sure to uncheck the box **Copy the main branch only** before clicking
-**Create fork**.
-
-For more details, see the example on `forking`_ from the GitHub documentation.
-
-Cloning the forked Minterpy repository
---------------------------------------
-
-You will then need to clone your fork to your local machine.
-Replace ``<placeholder>`` with your GitHub username and the name of your fork:
-
-.. code:: bash
-
-   git clone https://github.com/<your-username>/<your-fork> <your-local-directory>
-   cd <your-local-directory>
-   git remote add upstream https://github.com/casus/minterpy
-   git fetch upstream
-
-This creates the directory ``<your-local-directory>`` and connect your local
-repository to the upstream repository of Minterpy.
-
-For more details, see the example on `cloning`_ from the GitHub documentation.
-
-About the branching model
--------------------------
-
-We loosely follow the structure of `Gitflow`_ for our branching model.
-There are three types of branches in the Minterpy repository:
-
-1. `main branch`_: This branch contains only the latest release.
-   It includes fully tested, documented, and cleaned-up code.
-   The latest release of Minterpy is summarized in the `main branch CHANGELOG.md`_
-   (at the top with a numbered version).
-2. `dev branch`_: This branch stores the development version.
-   At any given time, the branch passes all the tests and contains a running
-   version of Minterpy, even if the code and documentation are not in a
-   "release state."
-   The latest state of development of Minterpy is summarized in the
-   `dev branch CHANGELOG.md`_ (under **Unreleased** section).
-3. Feature branches: All proposed changes to the code, be they new features,
-   enhancements, or bugfixes, happen on these branches.
-   These branches must always be branched out from the ``dev`` branch (and
-   not from ``main``).
-
-When you clone a forked Minterpy repository, you will be brought to the ``dev``
-branch by default.
-
-With your fork, you have the freedom to make any changes you like.
-However, when creating a merge (pull) request as a contribution, please note
-the following:
-
-- You need to create a merge (pull) request from a feature branch in your
-  repository to the ``dev`` branch of the Minterpy repository.
-- You cannot create a merge (pull) request to the ``main`` branch of
-  the Minterpy repository. The branch is used exclusively for release purposes.
[…]
