### 2025-04-25 93078ed
Revise landing page and displayed version information
The landing page has been revised for consistency
and the version that appear on the page has been
fixed to display the correct version.

This commit should resolve Issue #19.
