+++ pr-preview/42/_sources/api/polynomials/index.rst.txt
@@ -1,0 +1,14 @@
+====================
+minterpy.polynomials
+====================
+
+.. automodule:: minterpy.polynomials
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   Lagrange <lagrange>
+   Newton <newton>
+   Canonical <canonical>
+   Chebyshev <chebyshev>
