+++ pr-preview/42/_sources/api/internal/utils/poly-newton.rst.txt
@@ -1,0 +1,6 @@
+==================
+polynomials.newton
+==================
+
+.. automodule:: minterpy.utils.polynomials.newton
+   :members:
