+++ pr-preview/42/_sources/api/internal/utils/multi-index.rst.txt
@@ -1,0 +1,6 @@
+===========
+multi_index
+===========
+
+.. automodule:: minterpy.utils.multi_index
+   :members:
