+++ pr-preview/42/_sources/index.rst.txt
@@ -1,0 +1,123 @@
+:html_theme.sidebar_secondary.remove: true
+
+.. raw:: html
+
+   <div style="visibility: hidden;margin:-2.5em">
+
+======================
+Minterpy Documentation
+======================
+
+.. raw:: html
+
+   </div>
+
+.. image:: /assets/Wordmark-color.png
+
+Minterpy is an open-source Python package designed for constructing
+and manipulating multivariate interpolating polynomials with the goal of
+lifting the curse of dimensionality from interpolation tasks.
+
+.. grid:: auto
+   :margin: 0
+   :padding: 0
+   :gutter: 2
+
+   .. grid-item-card:: New to Minterpy?
+      :columns: 12 6 6 4
+      :class-card: sd-border-0
+      :class-footer: sd-border-0
+      :shadow: None
+
+      Discover what Minterpy and multivariate polynomials can do for you
+      through a series of practical tutorials.
+      +++
+      .. button-ref:: getting-started/index
+         :ref-type: myst
+         :class: landing-page-button
+         :expand:
+
+         :octicon:`rocket;2em` **Getting Started**
+
+   .. grid-item-card:: Familiar with Minterpy?
+      :columns: 12 6 6 4
+      :class-card: sd-border-0
+      :class-footer: sd-border-0
+      :shadow: None
+
+      Learn how to accomplish specific tasks with Minterpy.
+      +++
+      .. button-ref:: how-to/index
+         :ref-type: myst
+         :class: landing-page-button
+         :expand:
+
+         :octicon:`book;2em` **How-To Guides**
+
+   .. grid-item-card:: What's behind the scene?
+      :columns: 12 6 6 4
+      :class-card: sd-border-0
+      :class-footer: sd-border-0
+      :shadow: None
+
+      Gain a deeper understanding of the mathematics behind Minterpy.
+      +++
+      .. button-ref:: fundamentals/index
+         :ref-type: myst
+         :class: landing-page-button
+         :expand:
+
+         :octicon:`mortar-board;2em` **Fundamentals**
+
+   .. grid-item-card:: Need API reference?
+      :columns: 12 6 6 4
+      :class-card: sd-border-0
+      :class-footer: sd-border-0
+      :shadow: None
+
+      Explore the public API documentation of Minterpy.
+      +++
+      .. button-ref:: api/index
+         :ref-type: myst
+         :class: landing-page-button
+         :expand:
+
+         :octicon:`gear;2em` **API Reference**
+
+   .. grid-item-card:: Want to contribute?
+      :columns: 12 6 6 4
+      :class-card: sd-border-0
+      :class-footer: sd-border-0
+      :shadow: None
+
+      Get involved in the Minterpy development---contributions are
+      always welcome!
+      +++
+      .. button-ref:: contributors/index
+         :ref-type: myst
+         :class: landing-page-button
+         :expand:
+
+         :octicon:`git-pull-request;2em` **Contributors Guide**
+
+|
+
+Minterpy is being continuously extended and improved,
+with new functionalities added to address the bottlenecks
+involving interpolations in various computational tasks.
+
+|
+
+.. toctree::
+   :maxdepth: 3
+   :hidden:
+
+   Getting Started <getting-started/index>
+   How-to <how-to/index>
+   Fundamentals <fundamentals/index>
+   API Reference <api/index>
+   Contributors <contributors/index>
+
+----
+
+This is the documentation for Minterpy |version|, last updated on |today|.
