+++ pr-preview/42/_sources/api/internal/jit_compiled/multi-index.rst.txt
@@ -1,6 +1,0 @@
-===========
-multi_index
-===========
-
-.. automodule:: minterpy.jit_compiled.multi_index
-   :members:
