### 2025-04-26 #7702 @xuanxu
@kenohori can you confirm you also recommend the paper for publication?
