### 2025-04-28 minterpy-project/minterpy#14 @damar-wicaksono
Reviewer 1 has concluded their review rounds and [accepts](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2832296595) all the changes made.

All changes as of now are still in the `dev-joss` branch of the repository.
