### 2025-04-29 4352bce
Revise paper and documentation following JOSS review
The draft manuscript submitted to the Journal of Open Source Software (JOSS)
has been revised in accordance with the JOSS review process.
Additionally, some elements of the documentation
have been updated accordingly.

Specific changes made:

- Added commentary on the trade-off between accuracy and cost using global
  polynomials, especially in comparison to B-splines and local piecewise
  polynomials, which should resolve Issue #17.
- Updated README.md and CONTRIBUTING.md to fix broken links,
  update obsolete information, and improve clari
[…]
