+++ joss/paper.bib
@@ -64,19 +64,6 @@
   volume  = {5},
 }
 
-@InBook{Hecht2018,
-  author    = {Hecht, Michael and Sbalzarini, Ivo F.},
-  pages     = {53--75},
-  publisher = {Springer International Publishing},
-  title     = {Fast interpolation and fourier transform in high-dimensional spaces},
-  year      = {2018},
-  isbn      = {9783030011772},
-  month     = nov,
-  booktitle = {Intelligent Computing},
-  doi       = {10.1007/978-3-030-01177-2_5},
-  issn      = {2194-5365},
-}
-
 @Book{Trefethen2019,
   author    = {Trefethen, L. N.},
   publisher = {Society for Industrial and Applied Mathematics (SIAM)},
@@ -243,14 +230,6 @@
   publisher = {The Open Journal},
 }
 
-@Misc{Procida2024,
-  author = {Procida, Daniele},
-  note   = {Retrieved September 26, 2024},
-  title  = {{Diátaxis} documentation framework},
-  year   = {2024},
-  url    = {https://diataxis.fr/},
-}
-
 @Article{Trefethen2017a,
   author  = {Trefethen, L. N.},
   journal = {Proceedings of the American Mathematical Society},
@@ -287,6 +266,67 @@
   pages   = {176--200},
   volume  = {166},
   doi     = {10.1016/j.jat.2012.11.005},
+}
+
+@Book{Goldstine1977,
+  author    = {Goldstine, Herman H.},
+  publisher = {Springer New York},
+  title     = {A history of numerical analysis from the 16th through the 19th century},
+  year      = {1977},
+  isbn      = {9781468494723},
+  series    = {Studies in the History of Mathematics and Physical Sciences},
+  doi       = {10.1007/978-1-4684-9472-3},
+}
+
+@Article{Trefethen2016,
+  author    = {Trefethen, L. N.},
+  journal   = {Notices of the American Mathematical Society},
+  title     = {Inverse {Yogiisms}},
+  year      = {2016},
+  number    = {11},
+  pages     = {1281--1285},
+  volume    = {63},
+  doi       = {10.1090/noti1446},
+  publisher = {American Mathematical Society (AMS)},
+}
+
+@Article{Breuss2018,
+  author  = {Breuß, Michael and Kemm, Friedemann and Vogel, Oliver},
+  journal = {Kybernetika},
+  title   = {A numerical study of {Newton} interpolation with extremely high degrees},
+  year    = {2018},
+  pages   = {279--288},
+  doi     = {10.14736/kyb-2018-2-0279},
+}
+
+@Article{Reichel1990,
+  author  = {Reichel, Lothar},
+  journal = {BIT},
+  title   = {Newton interpolation at {Leja} points},
+  year    = {1990},
+  number  = {2},
+  pages   = {332--346},
+  volume  = {30},
+  doi     = {10.1007/bf02017352},
+}
+
+@Article{TalEzer1991,
+  author  = {Tal-Ezer, Hillel},
+  journal = {SIAM Journal on Scientific and Statistical Computing},
+  title   = {High degree polynomial interpolation in {Newton} form},
+  year    = {1991},
+  number  = {3},
+  pages   = {648--667},
+  volume  = {12},
+  doi     = {10.1137/0912034},
+}
+
+@Misc{Wicaksono2025,
+  author = {Wicaksono, Damar Canggih and Hernandez Acosta, Uwe and Thekke Veettil, Sachin Krishnan and Kissinger, Jannik and Hecht, Michael},
+  title  = {{Data to "Minterpy: Multivariate polynomial interpolation in Python"}},
+  year   = {2025},
[…]
