+++ joss/paper.md
@@ -42,21 +42,19 @@ bibliography: paper.bib
 
 # Summary
 
-Interpolation is the bottleneck in numerous computational tasks,
-appearing in function approximation, curve fitting, numerical integration,
-surface level-set methods, differential geometry,
-spectral methods, surrogate-model-based optimization,
-sensitivity analysis, uncertainty quantification,
-and beyond.
+Interpolation is essential in various computational tasks,
+including function approximation, curve fitting, numerical integration,
+differential geometry,
+spectral methods, optimization,
+and uncertainty quantification.
 
 Minterpy is an open-source Python package designed
 for multivariate polynomial interpolation.
-It provides a means of approximating a wide range of functions
-using numerically stable and accurate interpolating polynomials.
-Some of its key features are:
+It provides stable and accurate interpolating polynomials for approximating a wide range of functions.
+Key features include:
 
 - Polynomial interpolation on properly selected nodes and
-  polynomial regression on given arbitrary scattered nodes.
+  regression on arbitrary nodes.
 - Differentiation and integration operations on the polynomials.
 - Addition, subtraction, and multiplication operations
   on the polynomials.
@@ -67,318 +65,311 @@ commonly associated with interpolation tasks.
 
 # Statement of need
 
-Polynomial interpolation can be traced back to the works of
-Newton, Lagrange, and others [@Meijering2002],
-and its significance in mathematics and computing is well-established
-[@Cools2002;@Hecht2018;@Xiu2009].
-As a means of approximating functions, global polynomials---where a single polynomial
-is defined over the entire domain---offer several advantages
-that facilitate common numerical operations, such as
-differentiation, integration, addition, subtraction, or multiplication [@Trefethen2019].
+As a means of approximating functions, global polynomials---where a single
+polynomial is defined over the entire domain---offer several advantages.
+For sufficiently smooth functions, global polynomials can achieve
+high accuracy with a smaller number of data points (sampled over the entire domain)
+compared to local piecewise polynomials.
+Additionally, their relatively simple structure facilitates
+many common numerical operations.
+These operations include differentiation, integration, addition, subtraction,
+and multiplication [@Trefethen2019].
+
+The Stone-Weierstrass theorem establishes that any continuous function
+on a bounded domain in multiple dimensions can be approximated uniformly
+to arbitrary precision by multivariate global polynomials [@Branges1959].
+However, the theorem does not specify a concrete method for constructing
+such approximating polynomials.
+Various techniques can be employed to build approximating polynomials,
+such as least square approximations.
+Minterpy focuses on constructing approximating global polynomials using
+one of the earliest and most established methods: interpolation [@Goldstine1977].
 
 Polynomial interpolation is based on the principle that, in one dimension,
-one and only one polynomial $Q_{f, n}$ of degree $n$ can interpolate
-a function $f: \Omega \mapsto \mathbb{R}$ in a bounded domain $\Omega$
-with $n + 1$ distinct _unisolvent interpolation nodes (points)_ $P_n$ such that
-$$
+there exists a unique polynomial $Q_{f, n}$ of degree $n$ that interpolates
+a function $f: \Omega \to \mathbb{R}$ in a bounded domain
+$\Omega$ with $n + 1$ _distinct (unisolvent[^unisolvent]) interpolation nodes
+(or points)_ $P_n$ such that
+\begin{equation}
+\label{eq:interpolation-condition}
[…]

[…]
