### 2025-04-29 minterpy-project/minterpy#11 @damar-wicaksono
Reviewer has concluded their review rounds and [accepts](https://github.com/openjournals/joss-reviews/issues/7702#issuecomment-2835548315) all the changes made.

As of now, all changes are still on the `dev-joss` branch.
