+++ pr-preview/44/_sources/how-to/grid/grid-union.ipynb.txt
@@ -1,0 +1,266 @@
+{
+ "cells": [
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": "# Create the Union of `Grid` instances",
+   "id": "ebf1ed09291f1ec4"
+  },
+  {
+   "cell_type": "code",
+   "id": "initial_id",
+   "metadata": {
+    "collapsed": true
+   },
+   "source": [
+    "import matplotlib.pyplot as plt\n",
+    "import minterpy as mp\n",
+    "import numpy as np"
+   ],
+   "outputs": [],
+   "execution_count": null
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": [
+    "This guide demonstrates how to take the union of instances of `Grid` and the expected outcome of such an operation. The union between two instances of `Grid` may be obtained via the operator `|`.\n",
+    "\n",
+    "The union between two instances of `Grid` is another instance of `Grid` whose underlying multi-index set is the union of the index sets of the operands.\n",
+    "This `Grid` instance can support interpolation polynomials\n",
+    "that are the summed of polynomials that live on either `Grid` operands.\n",
+    "\n",
+    "To know more about the multi-index set union in Minterpy, see the {doc}`relevant example </how-to/multi-index-set/multi-index-set-union>`.\n",
+    "\n",
+    "When taking the union of two instances of `Grid` note the following caveats:\n",
+    "\n",
+    "- If a generating function is defined in both instances, then the instances can be multiplied only if the generating functions are compatible, i.e., they are the same function.\n",
+    "  defined in both instances.\n",
+    "- Without a generating function, two instances can be multiplied only if the generating\n",
+    "  points are compatible, i.e., they have the same values up to a common dimension (or columns)."
+   ],
+   "id": "8eee69b4e0b6b159"
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": [
+    "## Example: Instances with the same dimension\n",
+    "\n",
+    "As a motivating example, consider two two-dimensional interpolation grids:\n",
+    "\n",
+    "- The first grid has the multi-index set $\\{(0, 0), (1, 0), (0, 1), (1, 1), (0, 2), (1, 2) \\}$ with respect to $l_p$-degree $2.0$\n",
+    "- the second grid has a complete multi-index set with respect to a polynomial degree $2$ and $l_p$-degree $\\infty$.\n",
+    "\n",
+    "Both interpolation grids have points according to the Leja-ordered Chebyshev-Lobatto points (the default in Minterpy)."
+   ],
+   "id": "c35b5937514ac341"
+  },
+  {
+   "metadata": {},
+   "cell_type": "markdown",
+   "source": "First, create the instances of `Grid` according to the above specification:",
+   "id": "a957b70b132d8a14"
+  },
+  {
+   "metadata": {},
+   "cell_type": "code",
+   "source": [
+    "mi_1 = mp.MultiIndexSet(\n",
+    "    np.array([[0, 0], [1, 0], [0, 1], [1, 1], [0, 2], [0, 3]]),\n",
+    "    lp_degree=2.0,\n",
+    ")\n",
+    "grd_1 = mp.Grid(mi_1)\n",
+    "grd_2 = mp.Grid.from_degree(2, 2, np.inf)"
+   ],
+   "id": "db1a4a8e1f5a1401",
+   "outputs": [],
+   "execution_count": null
+  },
+  {
[…]
