+++ pr-preview/44/_sources/api/extras/regression/index.rst.txt
@@ -1,0 +1,12 @@
+==========
+regression
+==========
+
+.. automodule:: minterpy.extras.regression
+
+.. toctree::
+   :maxdepth: 1
+   :hidden:
+
+   Ordinary regression <ordinary_regression>
+   Regression abstract class <regression_abc>

[…]
