+++ joss/paper.bib
@@ -241,15 +241,6 @@
   doi     = {10.1090/proc/13623},
 }
 
-@Misc{Hecht2020,
-  author    = {Hecht, Michael and Gonciarz, Krzysztof and Michelfeit, Jannik and Sivkin, Vladimir and Sbalzarini, Ivo F.},
-  title     = {Multivariate interpolation in unisolvent nodes -- lifting the curse of dimensionality},
-  year      = {2020},
-  doi       = {10.48550/ARXIV.2010.10824},
-  keywords  = {Numerical Analysis (math.NA), FOS: Mathematics, FOS: Mathematics, Primary 65D15, 41A50, 41A63, 41A05, Secondary 41A25, 41A10},
-  publisher = {arXiv},
-}
-
 @Misc{Veettil2022,
   author = {Veettil, Sachin Krishnan Thekke and Zheng, Yuxi and Hernandez Acosta, Uwe and Wicaksono, Damar and Hecht, Michael},
   title  = {Multivariate polynomial regression of {Euclidean} degree extends the stability for fast approximations of {Trefethen} functions},
@@ -325,8 +316,17 @@
   author = {Wicaksono, Damar Canggih and Hernandez Acosta, Uwe and Thekke Veettil, Sachin Krishnan and Kissinger, Jannik and Hecht, Michael},
   title  = {{Data to "Minterpy: Multivariate polynomial interpolation in Python"}},
   year   = {2025},
-  doi    = {10.14278/rodare.3717},
-  url    = {https://doi.org/10.14278/rodare.3717},
+  doi    = {10.14278/rodare.3379},
+  url    = {https://doi.org/10.14278/rodare.3379},
+}
+
+@Misc{Hecht2025,
+  author        = {Michael Hecht and Phil-Alexander Hofmann and Damar Wicaksono and Uwe Hernandez Acosta and Krzysztof Gonciarz and Jannik Kissinger and Vladimir Sivkin and Ivo F. Sbalzarini},
+  title         = {Multivariate {Newton} interpolation in downward closed spaces reaches the optimal geometric approximation rates for {Bos}--{Levenberg}--{Trefethen} functions},
+  year          = {2025},
+  archiveprefix = {arXiv},
+  eprint        = {2504.17899},
+  url           = {https://arxiv.org/abs/2504.17899},
 }
 
 @Comment{jabref-meta: databaseType:bibtex;}
