### 2025-04-29 #7702 @damar-wicaksono
The reference methodological paper has been replaced with the most recent version, therefore...
