### 2025-04-29 minterpy-project/minterpy#12 @damar-wicaksono
The paper acceptance is imminent. The editor asks for the following tasks to be completed:
