### 2025-04-30 8c5f3c9
Fix typo in AUTHORS.md
