+++ pr-preview/44/_sources/api/interpolation.rst.txt
@@ -1,6 +1,0 @@
-======================
-minterpy.interpolation
-======================
-
-.. automodule:: minterpy.interpolation
-   :members:
