+++ pr-preview/46/_sources/fundamentals/dds.rst.txt
@@ -1,0 +1,212 @@
+================================================
+Multidimensional Divided Difference Scheme (DDS)
+================================================
+
+..
+    .. todo::
+
+       Multivariate DDS is central to the whole ``minterpy`` operation,
+       it is important somehow to *demistify* it.
+       In this page, all aspects related to the multivariate DDS should be defined
+       and explained from the data structure used, construction, and operation.
+
+       Provide illustrations and pictures for clarity and simple (numerical)
+       examples that one can do perhaps with hand to demistify it further.
+       ``minterpy`` "just" do it faster for larger problems.
+
+       .. todo::
+
+          Generate the links to
+
+          - downward closed multi-index sets
+          - polynomial space
+          - notion of unisolvence
+          - use latex label command for the Theorem enumeration etc
+
+Splitting Theorems
+==================
+
+The Multivariate DDS is based on two general splitting Theorems\ :footcite:`Hecht2020` that can be stated in a simplified version as:
+
+**Theorem 1 (Splitting of  unisolvent nodes)** Let :math:`A \subseteq \mathbb{N}^m` be a downward closed multi-index set,
+:math:`\Pi_A =\left<x^\alpha = x_1^{\alpha_1}\cdots x_m^{\alpha_m} : \alpha \in A\right>`
+be the polynomial space spanned by all monomials induced by :math:`A`,
+:math:`H = \{(x_1,\dots,x_m) \in \mathbb{R}^m : x_j = y\}, y \in [-1,1]` for some :math:`1\leq j \leq m` be a hyperplane parallel to the coordinate axis,
+:math:`Q_H= (x_j - q_j) \in \Pi_A` and :math:`P \subseteq \mathbb{N}^m`
+be a set of nodes such that:
+
+:math:`i)` There is no polynomial :math:`Q \in \Pi_A` being constant in normal direction to :math:`H`, i.e.,  :math:`\partial_{x_j}Q(x) =0`,
+and  :math:`Q` vanishes identically on :math:`P \cap H`, i.e., :math:`Q(p) =0\,,  \forall \,p \in  P \cap H`.
+
+:math:`ii)` There is no polynomial :math:`Q \in \Pi_A` with :math:`Q_H\cdot  Q \in \Pi_A`  and :math:`Q`
+vanishes identically on :math:`P \setminus H`, i.e, :math:`Q(p) =0\,,  \forall \, p \in  P \setminus H`.
+
+Then, the nodes :math:`P` are unisolvent with respect to :math:`\Pi_A`, i.e, the interpolant :math:`Q_{f,A}(p) = f(p) \,,  \forall \, p \in  P`  of any function :math:`f : \mathbb{R}^m \longrightarrow \mathbb{R}` is uniquely determined in :math:`\Pi_A`.
+
+Below, we show examples of unisolvent nodes in 2D for :math:`A= A_{2,3,1}`, generated by recursively applying Theorem 1. The three panels show examples for three different choices of the, in this case 1D, hyperplanes :math:`H_0,\ldots ,H_3`
+(solid lines). In the left panel, the hyperplanes are chosen arbitrarily. This starts by first choosing a hyperplane (line) :math:`H_0` and :math:`n+1=4`
+unisolvent nodes on :math:`H_0`. Then choose :math:`H_1 \not = H_0` and 3 unisolvent nodes on :math:`H_1 \setminus H_0`,
+and recursively continue until choosing 1 unisolvent node on :math:`H_3\setminus (H_0 \cup H_1 \cup H_2)`.
[…]
