+++ pr-preview/46/_sources/how-to/grid/grid-create-from-points.ipynb.txt
@@ -1,258 +1,0 @@
-{
- "cells": [
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": "# Create a Grid from an Array of Generating Points",
-   "id": "12e4f1e73ba755f"
-  },
-  {
-   "metadata": {},
-   "cell_type": "code",
-   "source": [
-    "import numpy as np\n",
-    "import minterpy as mp\n",
-    "import matplotlib.pyplot as plt"
-   ],
-   "id": "7937e29866199fe5",
-   "outputs": [],
-   "execution_count": null
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "Minterpy interpolating polynomials (i.e., in the Lagrange or Newton basis)\n",
-    "lives on a grid that holds the so-called _unisolvent nodes_.\n",
-    "An interpolating multi-dimensional polynomial of a given multi-index set of\n",
-    "exponents is uniquely\n",
-    "determined on the corresponding unisolvent nodes.\n",
-    "\n",
-    "An instance of `Grid` can be constructed via different constructors:\n",
-    "\n",
-    "- {py:meth}`.Grid.from_points`: create an instance with an array of generating points (this page)\n",
-    "- {py:meth}`.Grid.from_function`: create an instance with a generating function (see the {doc}`example <grid-create-from-function>`)\n",
-    "- {py:meth}`.Grid.from_degree`: create an instance with a complete multi-index set (see the {doc}`example <grid-create-from-degree>`)\n",
-    "- {py:meth}`.Grid.from_value_set`: create an instance based on an array of generating values (see the {doc}`example <grid-create-from-value-set>`)\n",
-    "- {py:class}`.Grid`: the default constructor (see the {doc}`example <grid-create-default-constructor>`)\n",
-    "\n",
-    "This guide provides an example on how to construct a `Grid` instance based\n",
-    "on a given array of generating points using the `from_points()` method."
-   ],
-   "id": "c3a659532557a7fc"
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "## About generating points\n",
-    "\n",
-    "Valid generating points to construct an instance of `Grid` are given as\n",
-    "a two-dimensional array of floats that must additionally satisfy\n",
-    "the following conditions:\n",
-    "\n",
-    "- the shape must be $(n + 1, m)$ where $n$ is the maximum polynomial degree\n",
-    "  across all dimensions and $m$ is the spatial dimension. Each column of the\n",
-    "  array is the one-dimensional interpolation points associated with that\n",
-    "  dimension.\n",
-    "- each column of the array must have unique values (no repeats).\n",
-    "- the number of rows of the array must be equal to or larger than the maximum\n",
-    "  exponents of the grid's multi-index set in any dimension."
-   ],
-   "id": "d9070443d6eacf2b"
-  },
-  {
-   "metadata": {},
-   "cell_type": "markdown",
-   "source": [
-    "## About `from_points()` factory method\n",
-    "\n",
-    "The method `from_points()` of the `Grid` class returns an instance of `Grid`\n",
-    "based on a given array of generating points.\n",
-    "\n",
-    "The method accepts two mandatory arguments, namely, the multi-index set\n",
-    "of exponents that defines the polynomials that grid can support,\n",
[…]

[…]
