+++ joss/paper.bib
@@ -321,7 +321,7 @@
 }
 
 @Misc{Hecht2025,
-  author        = {Michael Hecht and Phil-Alexander Hofmann and Damar Wicaksono and Uwe Hernandez Acosta and Krzysztof Gonciarz and Jannik Kissinger and Vladimir Sivkin and Ivo F. Sbalzarini},
+  author        = {Hecht, Michael and Hofmann, Phil-Alexander and Wicaksono, Damar and Hernandez Acosta, Uwe and Gonciarz, Krzysztof and Kissinger, Jannik and Sivkin, Vladimir and Sbalzarini, Ivo F.},
   title         = {Multivariate {Newton} interpolation in downward closed spaces reaches the optimal geometric approximation rates for {Bos}--{Levenberg}--{Trefethen} functions},
   year          = {2025},
   archiveprefix = {arXiv},
