+++ pr-preview/48/_sources/contributors/contrib-codebase/index.rst.txt
@@ -1,0 +1,25 @@
+============================
+Contributing to the Codebase
+============================
+
+If you're making changes to the Minterpy codebase (excluding its documentation),
+please review the following guidelines and conventions:
+
+- :doc:`code-style`: Ensure your code adheres to the established style guidelines.
+- :doc:`pre-commits`: Follow the pre-commit configuration to check the code
+  for syntax, formatting, and style errors.
+- :doc:`running-writing-tests`: Familiarize yourself with the testing framework
+  used in the Minterpy project and the convention used when writing new tests
+  to verify your changes.
+
+While the first two points may be discussed during a merge request,
+the last point is non-negotiable; all the tests--old and new--must pass.
+
+.. toctree::
+   :maxdepth: 2
+   :hidden:
+
+   code-style
+   pre-commits
+   running-writing-tests
+
