+++ pr-preview/48/_sources/api/polynomials/newton.rst.txt
@@ -1,0 +1,39 @@
+=================
+newton_polynomial
+=================
+
+.. currentmodule:: minterpy.polynomials.newton_polynomial
+
+.. automodule:: minterpy.polynomials.newton_polynomial
+
+.. autoclass:: NewtonPolynomial
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.polynomials.newton_polynomial.NewtonPolynomial
+      :properties:
+
+   .. rubric:: Methods
+
+   .. classautosummary:: minterpy.polynomials.newton_polynomial.NewtonPolynomial
+      :methods:
+
+
+----
+
+.. autofunction:: eval_newton
+
+.. autofunction:: add_newton
+
+.. autofunction:: mul_newton
+
+.. autofunction:: diff_newton
+
+.. autofunction:: partial_diff_newton
+
+.. autofunction:: integrate_over_newton
