+++ pr-preview/48/_sources/how-to/multi-index-set/multi-index-set-add.ipynb.txt
@@ -1,238 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "feeb9676-6a78-4dde-a023-8694cf99ae34",
-   "metadata": {},
-   "source": [
-    "# Add Exponents to a Multi-Index Set"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "61a9a739-bf5d-4707-ac2b-752c389e8ffa",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "5af0f904-1109-48ec-9d48-7c12c43bb91f",
-   "metadata": {},
-   "source": [
-    "An additional exponent or a set of exponents may be added to an already created `MultiIndexSet` instance. This guide shows you how."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "be4a3326-a453-4442-b806-09e0ac5b1453",
-   "metadata": {},
-   "source": [
-    "## Motivating example\n",
-    "\n",
-    "Suppose we have a complete multi-index set for multivariate polynomials of dimension $4$, degree $2$, and $l_p$-degree $1.0$:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "4b17f3ac-8ac5-4dd9-8700-9e5b1f7b53e5",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "mi = mp.MultiIndexSet.from_degree(4, 2, 1)"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "80c5a088-518f-43b3-8a68-4f891c56d066",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "print(mi)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "59544710-bbc3-4b38-983e-c6b919686098",
-   "metadata": {},
-   "source": [
-    "## Add an exponent to a `MultiIndexSet` instance"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "77509ab3-1307-43f0-b6a5-3189cfd47039",
-   "metadata": {},
-   "source": [
-    "The method `add_exponents()` inserts an exponent to a `MultiIndexSet` instance. The exponents must be given as a NumPy integer array of the correct dimension. \n",
-    "\n",
-    "Suppose now we'd like to add a new exponent element $\\left( 0, 0, 3, 0 \\right)$:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "a7766609-36a5-4ee0-beb2-e6d43a46bec2",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "new_exp = np.array([[0, 0, 3, 0]], dtype=int)\n",
-    "new_exp"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "be46c109-3055-477e-b94a-8349a3198b71",
-   "metadata": {},
-   "source": [
-    "Notice that the array is of shape $1 \\times 4$ because the dimension of the multi-index set is $4$ and one new element is to be added.\n",
-    "To add the exponent to the `MultiIndexSet` instance:"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "2f9ceb71-06d0-452e-80bb-513aa96b6cea",
-   "metadata": {},
-   "outputs": [],
-   "source": [
[…]
