+++ pr-preview/49/_sources/api/core/multi_index.rst.txt
@@ -1,0 +1,23 @@
+===========
+multi_index
+===========
+
+.. automodule:: minterpy.core.multi_index
+
+.. autoclass:: MultiIndexSet
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+   :exclude-members: __init__
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.core.multi_index.MultiIndexSet
+      :properties:
+
+   .. rubric:: Methods
+
+   .. classautosummary:: minterpy.core.multi_index.MultiIndexSet
+      :methods:
