+++ pr-preview/49/_sources/api/internal/utils/arrays.rst.txt
@@ -1,0 +1,6 @@
+======
+arrays
+======
+
+.. automodule:: minterpy.utils.arrays
+   :members:
