+++ pr-preview/49/_sources/api/internal/jit_compiled/newton-diff.rst.txt
@@ -1,0 +1,9 @@
+===========
+newton_diff
+===========
+
+.. currentmodule:: minterpy.jit_compiled.newton.diff
+
+.. automodule:: minterpy.jit_compiled.newton.diff
+   :members:
+   :ignore-module-all:
