+++ pr-preview/49/_sources/api/transformations/transformation-chebyshev.rst.txt
@@ -1,0 +1,43 @@
+=========
+chebyshev
+=========
+
+.. currentmodule:: minterpy.transformations.chebyshev
+
+.. automodule:: minterpy.transformations.chebyshev
+
+.. autoclass:: ChebyshevToLagrange
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.transformations.chebyshev.ChebyshevToLagrange
+      :properties:
+
+.. autoclass:: ChebyshevToNewton
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.transformations.chebyshev.ChebyshevToNewton
+      :properties:
+
+.. autoclass:: ChebyshevToCanonical
+   :members:
+   :show-inheritance:
+   :private-members:
+   :inherited-members:
+   :special-members:
+
+   .. rubric:: Properties
+
+   .. classautosummary:: minterpy.transformations.chebyshev.ChebyshevToCanonical
+      :properties:
