+++ pr-preview/49/_sources/getting-started/calculus-operations-with-polynomials.ipynb.txt
@@ -1,805 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "4f44cdf37aadcff5",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "# Basic Calculus Operations with Polynomials"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "c26c295879e98b28",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "So far, you've learned how to construct interpolating polynomials in both {doc}`one <1d-polynomial-interpolation>` and {doc}`multiple dimensions <md-polynomial-interpolation>`.\n",
-    "You've also explored the basic {doc}`arithmetic operations <arithmetic-operations-with-polynomials>` supported by Minterpy polynomials.\n",
-    "\n",
-    "In this in-depth tutorial, you'll learn about the basic calculus operations that can be performed with Minterpy polynomials."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "8fd381f9-68ea-4dd3-a055-85c8cadcd170",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "As usual, before you begin, you'll need to import the necessary packages to follow along with this guide."
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "1562e776-f727-4ab0-968e-8f798ec9729b",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np\n",
-    "import matplotlib.pyplot as plt"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "ea4cacd9-7afd-46af-bf13-582014249a75",
-   "metadata": {
-    "editable": true,
-    "slideshow": {
-     "slide_type": ""
-    },
-    "tags": []
-   },
-   "source": [
-    "## Motivating function\n",
-    "\n",
-    "Consider the following two-dimensional function:\n",
-    "\n",
-    "$$\n",
-    "\\begin{align}\n",
-    "\t\\mathcal{M}(\\boldsymbol{x}) = & 0.75 \\exp{\\left( -0.25 \\left( (x_1 - 2)^2 + (x_2 - 2)^2 \\right) \\right) } \\\\\n",
-    "                                  & + 0.75 \\exp{\\left( -1.00 \\left( \\frac{(x_1 + 1)^2}{49} + \\frac{(x_2 + 1)^2}{10} \\right) \\right)} \\\\\n",
-    "\t\t\t\t\t\t\t\t  & + 0.50 \\exp{\\left( -0.25 \\left( (x_1 - 7)^2 + (x_2 - 3)^2 \\right) \\right)} \\\\\n",
-    "\t\t\t\t\t\t\t\t  & - 0.20 \\exp{\\left( -1.00 \\left( (x_1 - 4)^2 + (x_2 - 7)^2 \\right) \\right)} \\\\\n",
-    "\\end{align}\n",
-    "$$\n",
-    "\n",
-    "where $\\boldsymbol{x} = (x_1, x_2) \\in [-1, 1]^2$."
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "429a1f91-4193-499e-a2c1-d9bc3f287306",
[…]
