+++ pr-preview/49/_sources/how-to/multi-index-set/multi-index-set-create.ipynb.txt
@@ -1,317 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "markdown",
-   "id": "d6a11f9a-d895-4cc7-befa-bac340b8915d",
-   "metadata": {},
-   "source": [
-    "# Create a Multi-Index Set"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": null,
-   "id": "b86f737a-bdc7-432e-ab3f-939491391220",
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "import minterpy as mp\n",
-    "import numpy as np"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "b81abe18-3ed4-4925-a09e-bdd95e4b0880",
-   "metadata": {},
-   "source": [
-    "A Minterpy polynomial is defined with respect to a multi-index set; the set defines the polynomial exponents across different spatial dimensions. This guide provides an example on how-to construct a multi-index set in Minterpy using different constructors.\n",
-    "\n",
-    "Currently, Minterpy supports the creation of a multi-index set via two constructors:\n",
-    "\n",
-    "- `minterpy.MultiIndexSet.from_degree()` is used to create a complete multi-index set with a given spatial dimension, polynomial degree, and $l_p$-degree\n",
-    "- `minterpy.MultiIndexSet()` (the default constructor) is used to create a multi-index set with a given set of exponents and $l_p$-degree"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "72d91dfd-ba93-48da-9de1-b7584c7a0428",
-   "metadata": {},
-   "source": [
-    "## From a given spatial dimension, polynomial degree, and $l_p$-degree"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "id": "55121107-73ca-422f-8afa-7fc6b1852c14",
-   "metadata": {},
-   "source": [
-    "The constructor `minterpy.MultiIndexSet.from_degree()` can be used to create a complete multi-index set for a given spatial dimension, degree, and $l_p$-degree.\n",
-    "The constructor takes the following (positional) arguments:\n",
-    "\n",
-    "- `spatial_dimension`: the number of spatial dimensions ($m$) of the multi-index set (must be an integer $> 0$)\n",
-    "- `poly_degree`: the polynomial degree ($n$) of the multi-index set (must be an integer $\\geq 0$)\n",
-    "- `lp_degree`: the degree $p$ of the $l_p$-norm of the multi-index set (optional but must be a positive float; if not given, a default value of $2.0$ is used)"
-   ]
-  },
-  {
-   "cell_type": "markdown",
-   "source": [
-    "```{note}\n",
-    "The set constructed via `from_degree()` constructor is _complete_ in the sense that it contains _all_ the exponents $\\boldsymbol{\\alpha} = \\left( \\alpha_1, \\ldots, \\alpha_m \\right) \\in \\mathbb{N}^m$ such that the $l_p$-norm $\\lVert \\boldsymbol{\\alpha} \\rVert_{p} = (\\alpha_1^{p} + \\ldots + \\alpha_m^{p})^{\\frac{1}{p}} \\leq n$ holds.\n",
-    "```"
-   ],
-   "metadata": {
-    "collapsed": false
-   },
-   "id": "57be42da-5e4d-4dba-995d-a2420b76091a"
-  },
-  {
-   "cell_type": "markdown",
-   "id": "28883cab-94ba-47ae-8b22-b1e05731a4bc",
-   "metadata": {},
-   "source": [
[…]
