+++ pr-preview/47/_sources/api/internal/jit_compiled/canonical.rst.txt
@@ -1,6 +1,0 @@
-=========
-canonical
-=========
-
-.. automodule:: minterpy.jit_compiled.canonical
-   :members:
