+++ stable/_sources/getting-started/1d-polynomial-interpolation.ipynb.txt
@@ -32,7 +32,7 @@
     "\n",
     "_But how to create such polynomials in the first place?_\n",
     "\n",
-    "This tutorial demonstrates how you can create such a polynomial in the most typical use case."
+    "This tutorial demonstrates how you can create such a polynomial from scratch in the most typical use case: interpolating a function."
    ]
   },
   {
@@ -129,8 +129,7 @@
    "source": [
     "## Motivating function\n",
     "\n",
-    "In Minterpy, the polynomials are typically (but not always) interpolating polynomials that approximate functions of interest.\n",
-    "In fact, this is one of the main objective of Minterpy: approximating a (multi-dimensional) function using a polynomial.\n",
+    "In Minterpy, polynomials are typically used to approximate functions of interest by interpolating them. This is actually one of the main goals of Minterpy: to approximate a (multidimensional) function using a polynomial.\n",
     "\n",
     "For now, however, consider the following one-dimensional function:\n",
     "\n",
@@ -180,7 +179,9 @@
     "slideshow": {
      "slide_type": ""
     },
-    "tags": []
+    "tags": [
+     "hide-input"
+    ]
    },
    "outputs": [],
    "source": [
@@ -202,7 +203,7 @@
     "tags": []
    },
    "source": [
-    "With the function of interest defined, you are now ready to create a polynomial that approximates it."
+    "Now that you've defined your function, you're ready to create a polynomial that approximates it."
    ]
   },
   {
@@ -225,9 +226,9 @@
     "3. **Evaluate the function of interest at the grid points**: Compute the function values at the grid points (also known as the unisolvent nodes)\n",
     "4. **Create a polynomial in the Lagrange basis**: Construct the interpolating polynomial in the Lagrange basis based on the evaluated points.\n",
     "\n",
-    "To perform additional operations with the interpolating polynomial, such as evaluating it a set of query points, follow this extra step:\n",
+    "To perform additional operations with the interpolating polynomial, such as evaluating it at a set of query points, follow this extra step:\n",
     "\n",
-    "5. **Transform the polynomial to another basis**: Change the basis of the interpolating polynomial from the Lagrange basis to another basis, preferably the Newton asis, for convenient evaluation and manipulation.\n",
+    "5. **Transform the polynomial to another basis**: Change the basis of the interpolating polynomial from the Lagrange basis to another basis, preferably the Newton basis, for convenient evaluation and manipulation.\n",
     "\n",
     "Let’s go through these steps one at a time."
    ]
@@ -314,7 +315,7 @@
    "source": [
     "### Construct an interpolation grid\n",
     "\n",
-    "An interpolation grid is where an interpolating polynomial lives on. The bases of interpolating polynomials, such as the Lagrange or the Newton basis, are defined with respect to these grids.\n",
+    "An interpolation grid is where an interpolating polynomial lives on. The bases of interpolating polynomials, such as the Lagrange or the Newton basis, are built with respect to these grids.\n",
     "\n",
     "In Minterpy, interpolation grids are represented by the {py:class}`Grid <.core.grid.Grid>` class.\n",
[…]
