+++ stable/_sources/getting-started/polynomial-regression.ipynb.txt
@@ -25,8 +25,8 @@
     "tags": []
    },
    "source": [
-    "The polynomials you've encountered so far in these tutorials are interpolating polynomials designed to approximate functions of interest based on values at specific, strategically chosen points known as the unisolvent nodes.\n",
-    "This also means that you have the flexibility to evaluate the function at these chosen points within its domain.\n",
+    "The polynomials you've worked with in these tutorials are designed to approximate functions of interest by using values at specific, strategically chosen points that form unisolvent nodes.\n",
+    "This implies that you have the flexibility to evaluate the function at these selected points within its domain.\n",
     "\n",
     "However, situations may arise where function values are available only at arbitrary, scattered points across the domain.\n",
     "In such cases, you cannot select the evaluation points freely.\n",
@@ -51,7 +51,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 1,
    "id": "e79dfcd0-a294-4cfe-9e67-7fb88eeb4579",
    "metadata": {
     "editable": true,
@@ -116,7 +116,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 2,
    "id": "5f86db01-c589-4155-8b05-576c046ac8d7",
    "metadata": {
     "editable": true,
@@ -153,7 +153,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 3,
    "id": "d173cf3e-6f8b-4941-8008-2139920c68f8",
    "metadata": {
     "editable": true,
@@ -185,7 +185,7 @@
   },
   {
    "cell_type": "code",
-   "execution_count": null,
+   "execution_count": 4,
[…]
