### 2025-04-30 #7702 @damar-wicaksono
@xuanxu,

- I've released a minor version of the package with the changes from this review (including minor documentation revision): [v0.3.1](https://github.com/minterpy-project/minterpy/releases/tag/v0.3.1)
- The release has been archived at [RODARE](https://doi.org/10.14278/rodare.3725)
- The DOI published by Zenodo is: 10.14278/rodare.3725

Please let me know if there's anything else I should do. Thanks a lot!
