### 2025-04-30 minterpy-project/minterpy#12 @damar-wicaksono
A new Minterpy release (v0.3.1) has been published both in [GitHub](https://github.com/minterpy-project/minterpy/releases/tag/v0.3.1) and [PyPI](https://pypi.org/project/minterpy/0.3.1/). Thanks to @szabo137, the entry at RODARE has been updated with [DOI: 10.14278/rodare.3725](https://doi.org/10.14278/rodare.3725).

We're now waiting for the handling editor to make the final (?) checks.
