### 2025-04-30 #7702 @damar-wicaksono
@xuanxu: For now, we've decided to keep the author order in the archive as it has been since the first entry, before I took on a more leading role in the development and co-maintenance of the package, and that's fine with me. Hopefully, this won't be an issue.
