### 2025-04-30 #7702 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
