### 2025-04-30 #7702 @xuanxu
Note: currently there is a problem with the Crossref servers causing doi verifications and depositing to fail.
Check https://status.crossref.org/ before moving forward
