### 2025-04-30 #7702 @danielskatz
👋 @damar-wicaksono - everything in the paper looks good to me. I'll accept it once Crossref is ready for us to publish it.  (thanks also to @xuanxu for making this easy for me)
