### 2025-05-01 #7702 @danielskatz
Congratulations to  @damar-wicaksono (Damar Wicaksono) and co-authors on your JOSS publication!!

JOSS completely depends on volunteers, and we couldn't do this without you
