### 2025-05-01 #7702 @damar-wicaksono
On behalf of the authors:

@xuanxu and @danielskatz: Thank you very much for your time and effort in editing and handling our submission to JOSS.

@sixpearls and @kenohori: We sincerely appreciate your thoughtful reviews and insightful feedback, which helped us improve the quality of our submission.
